[Decomp][Object] Cohesion split: extract update/values/position/chat/summon to dedicated TUs
