+++ src/game/Object/Object.cpp
@@ -187,3298 +187,7 @@ void Object::SetObjectScale(float newScale)
     SetFloatValue(OBJECT_FIELD_SCALE_X, newScale);
 }
 
-/**
- * @brief Force immediate update transmission to all viewers
- *
- * Sends all pending update changes immediately rather than waiting
- * for the next update tick. This is used for urgent updates that
- * must be visible immediately (e.g., combat state changes).
- *
- * The method builds update data for all nearby players and sends
- * it immediately, then removes the object from the pending update list.
- */
-void Object::SendForcedObjectUpdate()
-{
-    if (!m_inWorld || !m_objectUpdated)
-    {
-        return;
-    }
 
-    UpdateDataMapType update_players;
 
-    BuildUpdateData(update_players);
-    RemoveFromClientUpdateList();
 
-    WorldPacket packet;                                     // here we allocate a std::vector with a size of 0x10000
-    for (UpdateDataMapType::iterator iter = update_players.begin(); iter != update_players.end(); ++iter)
-    {
-        iter->second.BuildPacket(&packet);
-        iter->first->GetSession()->SendPacket(&packet);
-        packet.clear();                                     // clean the string
-    }
-}
 
-/**
- * @brief Build create update block for player
- * @param data Update data buffer
- * @param target Target player
- *
- * Builds the update packet data needed to create this object
- * for the specified player. Includes movement data and
- * all update field values.
- */
-void Object::BuildCreateUpdateBlockForPlayer(UpdateData* data, Player* target) const
-{
-    if (!target)
-    {
-        return;
-    }
-
-    uint8  updatetype   = UPDATETYPE_CREATE_OBJECT;
-    uint16 updateFlags  = m_updateFlag;
-
-    /** lower flag1 **/
-    if (target == this)                                     // building packet for yourself
-    {
-        updateFlags |= UPDATEFLAG_SELF;
-    }
-
-    if (m_itsNewObject)
-    {
-        switch (GetObjectGuid().GetHigh())
-        {
-            case HighGuid::HIGHGUID_DYNAMICOBJECT:
-            case HighGuid::HIGHGUID_CORPSE:
-            case HighGuid::HIGHGUID_PLAYER:
-            case HighGuid::HIGHGUID_UNIT:
-            case HighGuid::HIGHGUID_VEHICLE:
-            case HighGuid::HIGHGUID_GAMEOBJECT:
-                updatetype = UPDATETYPE_CREATE_OBJECT2;
-                break;
-
-            default:
-                break;
-        }
-    }
-
-    if (isType(TYPEMASK_UNIT))
-    {
-        if (((Unit*)this)->getVictim())
-        {
-            updateFlags |= UPDATEFLAG_HAS_ATTACKING_TARGET;
-        }
-    }
-
-    // DEBUG_LOG("BuildCreateUpdate: update-type: %u, object-type: %u got updateFlags: %X", updatetype, m_objectTypeId, updateFlags);
-
-    ByteBuffer& buf = data->GetBuffer();
-    buf << uint8(updatetype);
-    buf << GetPackGUID();
-    buf << uint8(m_objectTypeId);
-
-    BuildMovementUpdate(&buf, updateFlags);
-
-    UpdateMask updateMask;
-    updateMask.SetCount(m_valuesCount);
-    _SetCreateBits(&updateMask, target);
-    BuildValuesUpdate(updatetype, &buf, &updateMask, target);
-    data->AddUpdateBlock();
-}
-
-/**
- * @brief Send create update to player
- * @param player Target player
- *
- * Sends the create update packet to the specified player,
- * causing the object to appear in their game world.
- */
-void Object::SendCreateUpdateToPlayer(Player* player)
-{
-    // send create update to player
-    UpdateData upd(player->GetMapId());
-    WorldPacket packet;
-
-    BuildCreateUpdateBlockForPlayer(&upd, player);
-    upd.BuildPacket(&packet);
-    player->GetSession()->SendPacket(&packet);
-}
-
-/**
- * @brief Build values update block for player
- * @param data Update data buffer
- * @param target Target player
- *
- * Builds the update packet data for changed field values
- * to send to the specified player.
- */
-void Object::BuildValuesUpdateBlockForPlayer(UpdateData* data, Player* target) const
-{
-    ByteBuffer& buf = data->GetBuffer();
-
-    buf << uint8(UPDATETYPE_VALUES);
-    buf << GetPackGUID();
-
-    UpdateMask updateMask;
-    updateMask.SetCount(m_valuesCount);
-
-    _SetUpdateBits(&updateMask, target);
-    BuildValuesUpdate(UPDATETYPE_VALUES, &buf, &updateMask, target);
-
-    data->AddUpdateBlock();
-}
-
-/**
- * @brief Build out of range update block
- * @param data Update data buffer
- *
- * Adds this object's GUID to the out-of-range list,
- * indicating it should be removed from the client's view.
- */
-void Object::BuildOutOfRangeUpdateBlock(UpdateData* data) const
-{
-    data->AddOutOfRangeGUID(GetObjectGuid());
-}
-
-/**
- * @brief Destroy object for player
- * @param target Target player
- *
- * Sends a destroy packet to the specified player,
- * removing this object from their game world.
- */
-void Object::DestroyForPlayer(Player* target, bool anim) const
-{
-    MANGOS_ASSERT(target);
-
-    WorldPacket data(SMSG_DESTROY_OBJECT, 9);
-    data << GetObjectGuid();
-    data << uint8(anim ? 1 : 0);                            // WotLK (bool), may be despawn animation
-    target->GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Build movement update block
- * @param data Byte buffer to write to
- * @param updateFlags Update flags
- *
- * Builds the movement data portion of the update packet.
- * Includes position, orientation, movement flags, and speeds
- * for living objects, or just position for static objects.
- */
-void Object::BuildMovementUpdate(ByteBuffer* data, uint16 updateFlags) const
-{
-    ObjectGuid Guid = GetObjectGuid();
-
-    data->WriteBit(false);
-    data->WriteBit(false);
-    data->WriteBit(updateFlags & UPDATEFLAG_ROTATION);
-    data->WriteBit(updateFlags & UPDATEFLAG_ANIM_KITS);               // AnimKits
-    data->WriteBit(updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET);
-    data->WriteBit(updateFlags & UPDATEFLAG_SELF);
-    data->WriteBit(updateFlags & UPDATEFLAG_VEHICLE);
-    data->WriteBit(updateFlags & UPDATEFLAG_LIVING);
-    data->WriteBits(0, 24);                                     // Byte Counter
-    data->WriteBit(false);
-    data->WriteBit(updateFlags & UPDATEFLAG_POSITION);                // flags & UPDATEFLAG_HAS_POSITION Game Object Position
-    data->WriteBit(updateFlags & UPDATEFLAG_HAS_POSITION);            // Stationary Position
-    data->WriteBit(updateFlags & UPDATEFLAG_TRANSPORT_ARR);
-    data->WriteBit(false);
-    data->WriteBit(updateFlags & UPDATEFLAG_TRANSPORT);
-
-    bool hasTransport = false,
-        isSplineEnabled = false,
-        hasPitch = false,
-        hasFallData = false,
-        hasFallDirection = false,
-        hasElevation = false,
-        hasOrientation = !isType(TYPEMASK_ITEM),
-        hasTimeStamp = true,
-        hasTransportTime2 = false,
-        hasTransportTime3 = false;
-
-    if (isType(TYPEMASK_UNIT))
-    {
-        Unit const* unit = (Unit const*)this;
-        hasTransport = !unit->m_movementInfo.GetTransportGuid().IsEmpty();
-        isSplineEnabled = unit->IsSplineEnabled();
-
-        if (GetTypeId() == TYPEID_PLAYER)
-        {
-            // use flags received from client as they are more correct
-            hasPitch = unit->m_movementInfo.GetStatusInfo().hasPitch;
-            hasFallData = unit->m_movementInfo.GetStatusInfo().hasFallData;
-            hasFallDirection = unit->m_movementInfo.GetStatusInfo().hasFallDirection;
-            hasElevation = unit->m_movementInfo.GetStatusInfo().hasSplineElevation;
-            hasTransportTime2 = unit->m_movementInfo.GetStatusInfo().hasTransportTime2;
-            hasTransportTime3 = unit->m_movementInfo.GetStatusInfo().hasTransportTime3;
-        }
-        else
-        {
-            hasPitch = unit->m_movementInfo.HasMovementFlag(MovementFlags(MOVEFLAG_SWIMMING | MOVEFLAG_FLYING)) ||
-                            unit->m_movementInfo.HasMovementFlag2(MOVEFLAG2_ALLOW_PITCHING);
-            hasFallData = unit->m_movementInfo.HasMovementFlag2(MOVEFLAG2_INTERP_TURNING);
-            hasFallDirection = unit->m_movementInfo.HasMovementFlag(MOVEFLAG_FALLING);
-            hasElevation = unit->m_movementInfo.HasMovementFlag(MOVEFLAG_SPLINE_ELEVATION);
-        }
-    }
-
-    if (updateFlags & UPDATEFLAG_LIVING)
-    {
-        Unit const* unit = (Unit const*)this;
-
-        data->WriteBit(!unit->m_movementInfo.GetMovementFlags());
-        data->WriteBit(!hasOrientation);
-
-        data->WriteGuidMask<7, 3, 2>(Guid);
-
-        if (unit->m_movementInfo.GetMovementFlags())
-        {
-            data->WriteBits(unit->m_movementInfo.GetMovementFlags(), 30);
-        }
-
-        data->WriteBit(false);
-        data->WriteBit(!hasPitch);
-        data->WriteBit(isSplineEnabled);
-        data->WriteBit(hasFallData);
-        data->WriteBit(!hasElevation);
-        data->WriteGuidMask<5>(Guid);
-        data->WriteBit(hasTransport);
-        data->WriteBit(!hasTimeStamp);
-
-        if (hasTransport)
-        {
-            ObjectGuid tGuid = unit->m_movementInfo.GetTransportGuid();
-
-            data->WriteGuidMask<1>(tGuid);
-            data->WriteBit(hasTransportTime2);
-            data->WriteGuidMask<4, 0, 6>(tGuid);
-            data->WriteBit(hasTransportTime3);
-            data->WriteGuidMask<7, 5, 3, 2>(tGuid);
-        }
-
-        data->WriteGuidMask<4>(Guid);
-
-        if (isSplineEnabled)
-        {
-            Movement::PacketBuilder::WriteCreateBits(*unit->movespline, *data);
-        }
-
-        data->WriteGuidMask<6>(Guid);
-
-        if (hasFallData)
-        {
-            data->WriteBit(hasFallDirection);
-        }
-
-        data->WriteGuidMask<0, 1>(Guid);
-        data->WriteBit(false);    // Unknown 4.3.3
-        data->WriteBit(!unit->m_movementInfo.GetMovementFlags2());
-
-        if (unit->m_movementInfo.GetMovementFlags2())
-        {
-            data->WriteBits(unit->m_movementInfo.GetMovementFlags2(), 12);
-        }
-    }
-
-    // used only with GO's, placeholder
-    if (updateFlags & UPDATEFLAG_POSITION)
-    {
-        ObjectGuid transGuid;
-        data->WriteGuidMask<5>(transGuid);
-        data->WriteBit(hasTransportTime3);
-        data->WriteGuidMask<0, 3, 6, 1, 4, 2>(transGuid);
-        data->WriteBit(hasTransportTime2);
-        data->WriteGuidMask<7>(transGuid);
-    }
-
-    if (updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET)
-    {
-        ObjectGuid guid;
-        if (Unit* victim = ((Unit*)this)->getVictim())
-        {
-            guid = victim->GetObjectGuid();
-        }
-
-        data->WriteGuidMask<2, 7, 0, 4, 5, 6, 1, 3>(guid);
-    }
-
-    if (updateFlags & UPDATEFLAG_ANIM_KITS)
-    {
-        data->WriteBit(true);   // hasAnimKit0 == false
-        data->WriteBit(true);   // hasAnimKit1 == false
-        data->WriteBit(true);   // hasAnimKit2 == false
-    }
-
-    data->FlushBits();
-
-    if (updateFlags & UPDATEFLAG_LIVING)
-    {
-        Unit const* unit = (Unit const*)this;
-
-        data->WriteGuidBytes<4>(Guid);
-
-        *data << float(unit->GetSpeed(MOVE_RUN_BACK));
-
-        if (hasFallData)
-        {
-            if (hasFallDirection)
-            {
-                // 15595 client reads horizontal speed, then jump direction sin/cos
-                *data << float(unit->m_movementInfo.GetJumpInfo().xyspeed);
-                *data << float(unit->m_movementInfo.GetJumpInfo().sinAngle);
-                *data << float(unit->m_movementInfo.GetJumpInfo().cosAngle);
-            }
-
-            *data << uint32(unit->m_movementInfo.GetFallTime());
-            *data << float(unit->m_movementInfo.GetJumpInfo().velocity);
-        }
-
-        *data << float(unit->GetSpeed(MOVE_SWIM_BACK));
-
-        if (hasElevation)
-        {
-            *data << float(unit->m_movementInfo.GetSplineElevation());
-        }
-
-        if (isSplineEnabled)
-        {
-            Movement::PacketBuilder::WriteCreateBytes(*unit->movespline, *data);
-        }
-
-        *data << float(unit->GetPositionZ());
-        data->WriteGuidBytes<5>(Guid);
-
-        if (hasTransport)
-        {
-            ObjectGuid tGuid = unit->m_movementInfo.GetTransportGuid();
-
-            data->WriteGuidBytes<5, 7>(tGuid);
-            *data << uint32(unit->m_movementInfo.GetTransportTime());
-            *data << float(NormalizeOrientation(unit->m_movementInfo.GetTransportPos()->o));
-
-            if (hasTransportTime2)
-            {
-                *data << uint32(unit->m_movementInfo.GetTransportTime2());
-            }
-
-            *data << float(unit->m_movementInfo.GetTransportPos()->y);
-            *data << float(unit->m_movementInfo.GetTransportPos()->x);
-            data->WriteGuidBytes<3>(tGuid);
-            *data << float(unit->m_movementInfo.GetTransportPos()->z);
-            data->WriteGuidBytes<0>(tGuid);
-
-            if (hasTransportTime3)
-            {
-                *data << uint32(unit->m_movementInfo.GetFallTime());
-            }
-
-            *data << int8(unit->m_movementInfo.GetTransportSeat());
-            data->WriteGuidBytes<1, 6, 2, 4>(tGuid);
-        }
-
-        *data << float(unit->GetPositionX());
-        *data << float(unit->GetSpeed(MOVE_PITCH_RATE));
-        data->WriteGuidBytes<3, 0>(Guid);
-        *data << float(unit->GetSpeed(MOVE_SWIM));
-        *data << float(unit->GetPositionY());
-        data->WriteGuidBytes<7, 1, 2>(Guid);
-        *data << float(unit->GetSpeed(MOVE_WALK));
-
-        *data << uint32(GameTime::GetGameTimeMS());
-
-        *data << float(unit->GetSpeed(MOVE_FLIGHT_BACK));
-        data->WriteGuidBytes<6>(Guid);
-        *data << float(unit->GetSpeed(MOVE_TURN_RATE));
-
-        if (hasOrientation)
-        {
-            *data << float(NormalizeOrientation(unit->GetOrientation()));
-        }
-
-        *data << float(unit->GetSpeed(MOVE_RUN));
-
-        if (hasPitch)
-        {
-            *data << float(unit->m_movementInfo.GetPitch());
-        }
-
-        *data << float(unit->GetSpeed(MOVE_FLIGHT));
-    }
-
-    if (updateFlags & UPDATEFLAG_VEHICLE)
-    {
-        *data << float(NormalizeOrientation(((WorldObject*)this)->GetOrientation()));
-        *data << uint32(((Unit*)this)->GetVehicleInfo()->GetVehicleEntry()->m_ID); // vehicle id
-    }
-
-    // used only with GO's, placeholder
-    if (updateFlags & UPDATEFLAG_POSITION)
-    {
-        ObjectGuid transGuid;
-
-        data->WriteGuidBytes<0, 5>(transGuid);
-        if (hasTransportTime3)
-        {
-            *data << uint32(0);
-        }
-
-        data->WriteGuidBytes<3>(transGuid);
-        *data << float(0.0f);   // x offset
-        data->WriteGuidBytes<4, 6, 1>(transGuid);
-        *data << uint32(0);     // transport time
-        *data << float(0.0f);   // y offset
-        data->WriteGuidBytes<2, 7>(transGuid);
-        *data << float(0.0f);   // z offset
-        *data << int8(-1);      // transport seat
-        *data << float(0.0f);   // o offset
-
-        if (hasTransportTime2)
-        {
-            *data << uint32(0);
-        }
-    }
-
-    if (updateFlags & UPDATEFLAG_ROTATION)
-    {
-        *data << int64(((GameObject*)this)->GetPackedWorldRotation());
-    }
-
-    if (updateFlags & UPDATEFLAG_TRANSPORT_ARR)
-    {
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << uint8(0);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-        *data << float(0.0f);
-    }
-
-    if (updateFlags & UPDATEFLAG_HAS_POSITION)
-    {
-        *data << float(NormalizeOrientation(((WorldObject*)this)->GetOrientation()));
-        *data << float(((WorldObject*)this)->GetPositionX());
-        *data << float(((WorldObject*)this)->GetPositionY());
-        *data << float(((WorldObject*)this)->GetPositionZ());
-    }
-
-    if (updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET)
-    {
-        ObjectGuid guid;
-        if (Unit* victim = ((Unit*)this)->getVictim())
-        {
-            guid = victim->GetObjectGuid();
-        }
-
-        data->WriteGuidBytes<4, 0, 3, 5, 7, 6, 2, 1>(guid);
-    }
-
-    if (updateFlags & UPDATEFLAG_TRANSPORT)
-    {
-        *data << uint32(GameTime::GetGameTimeMS());           // ms time
-    }
-}
-
-/**
- * @brief Build values update data
- * @param updatetype Update type (create or values)
- * @param data Byte buffer to write to
- * @param updateMask Update mask indicating which fields changed
- * @param target Target player
- *
- * Builds the actual field value data for the update packet.
- * Handles special cases for gameobjects and units.
- */
-void Object::BuildValuesUpdate(uint8 updatetype, ByteBuffer* data, UpdateMask* updateMask, Player* target) const
-{
-    if (!target)
-    {
-        return;
-    }
-
-    uint32 valuesCount = m_valuesCount;
-    if (GetTypeId() == TYPEID_PLAYER && target != this)
-    {
-        valuesCount = PLAYER_END_NOT_SELF;
-    }
-
-    bool IsActivateToQuest = false;
-    bool IsPerCasterAuraState = false;
-
-    if (updatetype == UPDATETYPE_CREATE_OBJECT || updatetype == UPDATETYPE_CREATE_OBJECT2)
-    {
-        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
-        {
-            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
-            {
-                IsActivateToQuest = true;
-            }
-
-            updateMask->SetBit(GAMEOBJECT_DYNAMIC);
-        }
-        else if (isType(TYPEMASK_UNIT))
-        {
-            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
-            {
-                IsPerCasterAuraState = true;
-                updateMask->SetBit(UNIT_FIELD_AURASTATE);
-            }
-        }
-    }
-    else                                                    // case UPDATETYPE_VALUES
-    {
-        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
-        {
-            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
-            {
-                IsActivateToQuest = true;
-            }
-
-            updateMask->SetBit(GAMEOBJECT_DYNAMIC);
-            updateMask->SetBit(GAMEOBJECT_BYTES_1);         // why do we need this here?
-        }
-        else if (isType(TYPEMASK_UNIT))
-        {
-            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
-            {
-                IsPerCasterAuraState = true;
-                updateMask->SetBit(UNIT_FIELD_AURASTATE);
-            }
-        }
-    }
-
-    MANGOS_ASSERT(updateMask && updateMask->GetCount() == m_valuesCount);
-
-    *data << (uint8)updateMask->GetBlockCount();
-    data->append(updateMask->GetMask(), updateMask->GetLength());
-
-    // 2 specialized loops for speed optimization in non-unit case
-    if (isType(TYPEMASK_UNIT))                              // unit (creature/player) case
-    {
-        for (uint16 index = 0; index < valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                if (index == UNIT_NPC_FLAGS)
-                {
-                    uint32 appendValue = m_uint32Values[index];
-
-                    if (GetTypeId() == TYPEID_UNIT)
-                    {
-                        if (!target->canSeeSpellClickOn((Creature*)this))
-                        {
-                            appendValue &= ~UNIT_NPC_FLAG_SPELLCLICK;
-                        }
-
-                        if (appendValue & UNIT_NPC_FLAG_TRAINER)
-                        {
-                            if (!((Creature*)this)->IsTrainerOf(target, false))
-                            {
-                                appendValue &= ~(UNIT_NPC_FLAG_TRAINER | UNIT_NPC_FLAG_TRAINER_CLASS | UNIT_NPC_FLAG_TRAINER_PROFESSION);
-                            }
-                        }
-
-                        if (appendValue & UNIT_NPC_FLAG_STABLEMASTER)
-                        {
-                            if (target->getClass() != CLASS_HUNTER)
-                            {
-                                appendValue &= ~UNIT_NPC_FLAG_STABLEMASTER;
-                            }
-                        }
-                    }
-
-                    *data << uint32(appendValue);
-                }
-                else if (index == UNIT_FIELD_AURASTATE)
-                {
-                    if (IsPerCasterAuraState)
-                    {
-                        // IsPerCasterAuraState set if related pet caster aura state set already
-                        if (((Unit*)this)->HasAuraStateForCaster(AURA_STATE_CONFLAGRATE, target->GetObjectGuid()))
-                        {
-                            *data << m_uint32Values[index];
-                        }
-                        else
-                        {
-                            *data << (m_uint32Values[index] & ~(1 << (AURA_STATE_CONFLAGRATE - 1)));
-                        }
-                    }
-                    else
-                    {
-                        *data << m_uint32Values[index];
-                    }
-                }
-                // FIXME: Some values at server stored in float format but must be sent to client in uint32 format
-                else if (index >= UNIT_FIELD_BASEATTACKTIME && index <= UNIT_FIELD_RANGEDATTACKTIME)
-                {
-                    // convert from float to uint32 and send
-                    *data << uint32(m_floatValues[index] < 0 ? 0 : m_floatValues[index]);
-                }
-
-                // there are some float values which may be negative or can't get negative due to other checks
-                else if ((index >= UNIT_FIELD_NEGSTAT0 && index <= UNIT_FIELD_NEGSTAT4) ||
-                         (index >= UNIT_FIELD_RESISTANCEBUFFMODSPOSITIVE  && index <= (UNIT_FIELD_RESISTANCEBUFFMODSPOSITIVE + 6)) ||
-                         (index >= UNIT_FIELD_RESISTANCEBUFFMODSNEGATIVE  && index <= (UNIT_FIELD_RESISTANCEBUFFMODSNEGATIVE + 6)) ||
-                         (index >= UNIT_FIELD_POSSTAT0 && index <= UNIT_FIELD_POSSTAT4))
-                {
-                    *data << uint32(m_floatValues[index]);
-                }
-
-                // Gamemasters should be always able to select units - remove not selectable flag
-                else if (index == UNIT_FIELD_FLAGS && target->isGameMaster())
-                {
-                    *data << (m_uint32Values[index] & ~UNIT_FLAG_NOT_SELECTABLE);
-                }
-                /* Hide loot animation for players that aren't permitted to loot the corpse */
-                else if (index == UNIT_DYNAMIC_FLAGS && GetTypeId() == TYPEID_UNIT)
-                {
-                    uint32 send_value = m_uint32Values[index];
-
-                    /* Initiate pointer to creature so we can check loot */
-                    if (Creature* my_creature = (Creature*)this)
-                        /* If the creature is NOT fully looted */
-                        if (!my_creature->loot.isLooted())
-                            /* If the lootable flag is NOT set */
-                            if (!(send_value & UNIT_DYNFLAG_LOOTABLE))
-                            {
-                                /* Update it on the creature */
-                                my_creature->SetFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_LOOTABLE);
-                                /* Update it in the packet */
-                                send_value = send_value | UNIT_DYNFLAG_LOOTABLE;
-                            }
-
-                    /* If we're not allowed to loot the target, destroy the lootable flag */
-                    if (!target->isAllowedToLoot((Creature*)this))
-                        if (send_value & UNIT_DYNFLAG_LOOTABLE)
-                        {
-                            send_value = send_value & ~UNIT_DYNFLAG_LOOTABLE;
-                        }
-
-                    /* If we are allowed to loot it and mob is tapped by us, destroy the tapped flag */
-                    bool is_tapped = target->IsTappedByMeOrMyGroup((Creature*)this);
-
-                    /* If the creature has tapped flag but is tapped by us, remove the flag */
-                    if (send_value & UNIT_DYNFLAG_TAPPED && is_tapped)
-                    {
-                        send_value = send_value & ~UNIT_DYNFLAG_TAPPED;
-                    }
-
-                    *data << send_value;
-                }
-                else
-                {
-                    // send in current format (float as float, uint32 as uint32)
-                    *data << m_uint32Values[index];
-                }
-            }
-        }
-    }
-    else if (isType(TYPEMASK_GAMEOBJECT))                   // gameobject case
-    {
-        for (uint16 index = 0; index < valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                // send in current format (float as float, uint32 as uint32)
-                if (index == GAMEOBJECT_DYNAMIC)
-                {
-                    // GAMEOBJECT_TYPE_DUNGEON_DIFFICULTY can have lo flag = 2
-                    //      most likely related to "can enter map" and then should be 0 if can not enter
-
-                    if (IsActivateToQuest)
-                    {
-                        switch (((GameObject*)this)->GetGoType())
-                        {
-                            case GAMEOBJECT_TYPE_QUESTGIVER:
-                                // GO also seen with GO_DYNFLAG_LO_SPARKLE explicit, relation/reason unclear (192861)
-                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE);
-                                *data << uint16(-1);
-                                break;
-                            case GAMEOBJECT_TYPE_CHEST:
-                            case GAMEOBJECT_TYPE_GENERIC:
-                            case GAMEOBJECT_TYPE_SPELL_FOCUS:
-                            case GAMEOBJECT_TYPE_GOOBER:
-                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE | GO_DYNFLAG_LO_SPARKLE);
-                                *data << uint16(-1);
-                                break;
-                            default:
-                                // unknown, not happen.
-                                *data << uint16(0);
-                                *data << uint16(-1);
-                                break;
-                        }
-                    }
-                    else
-                    {
-                        // disable quest object
-                        *data << uint16(0);
-                        *data << uint16(-1);
-                    }
-                }
-                else if (index == GAMEOBJECT_BYTES_1)
-                {
-                    if (((GameObject*)this)->GetGOInfo()->type == GAMEOBJECT_TYPE_TRANSPORT)
-                    {
-                        *data << uint32(m_uint32Values[index] | GO_STATE_TRANSPORT_SPEC);
-                    }
-                    else
-                    {
-                        *data << uint32(m_uint32Values[index]);
-                    }
-                }
-                else
-                {
-                    *data << m_uint32Values[index];          // other cases
-                }
-            }
-        }
-    }
-    else                                                    // other objects case (no special index checks)
-    {
-        for (uint16 index = 0; index < valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                // send in current format (float as float, uint32 as uint32)
-                *data << m_uint32Values[index];
-            }
-        }
-    }
-}
-
-/**
- * @brief Clear update mask
- * @param remove If true, remove from client update list
- *
- * Clears all changed value flags and optionally removes
- * the object from the pending update list.
- */
-void Object::ClearUpdateMask(bool remove)
-{
-    if (m_uint32Values)
-    {
-        for (uint16 index = 0; index < m_valuesCount; ++index)
-        {
-            m_changedValues[index] = false;
-        }
-    }
-
-    if (m_objectUpdated)
-    {
-        if (remove)
-        {
-            RemoveFromClientUpdateList();
-        }
-        m_objectUpdated = false;
-    }
-}
-
-/**
- * @brief Load values from data string
- * @param data Data string to load from
- * @return True if successful
- *
- * Loads update field values from a character data string.
- * Used when loading objects from database.
- */
-bool Object::LoadValues(const char* data)
-{
-    if (!m_uint32Values)
-    {
-        _InitValues();
-    }
-
-    Tokens tokens = StrSplit(data, " ");
-
-    if (tokens.size() != m_valuesCount)
-    {
-        return false;
-    }
-
-    Tokens::iterator iter;
-    int index;
-    for (iter = tokens.begin(), index = 0; index < m_valuesCount; ++iter, ++index)
-    {
-        m_uint32Values[index] = std::stoul((*iter).c_str());
-    }
-
-    return true;
-}
-
-/**
- * @brief Set update bits in mask
- * @param updateMask Update mask to modify
- * @param target Target player (unused)
- *
- * Sets bits in the update mask for all fields that have changed.
- */
-void Object::_SetUpdateBits(UpdateMask* updateMask, Player* target) const
-{
-    uint32 valuesCount = m_valuesCount;
-    if (GetTypeId() == TYPEID_PLAYER && target != this)
-    {
-        valuesCount = PLAYER_END_NOT_SELF;
-    }
-
-    for (uint16 index = 0; index < valuesCount; ++index )
-        if (m_changedValues[index])
-        {
-            updateMask->SetBit(index);
-        }
-}
-
-/**
- * @brief Set create bits in mask
- * @param updateMask Update mask to modify
- * @param target Target player (unused)
- *
- * Sets bits in the update mask for all non-zero fields.
- * Used when creating a new object for a player.
- */
-void Object::_SetCreateBits(UpdateMask* updateMask, Player* target) const
-{
-    uint32 valuesCount = m_valuesCount;
-    if (GetTypeId() == TYPEID_PLAYER && target != this)
-    {
-        valuesCount = PLAYER_END_NOT_SELF;
-    }
-
-    for (uint16 index = 0; index < valuesCount; ++index)
-        if (GetUInt32Value(index) != 0)
-        {
-            updateMask->SetBit(index);
-        }
-}
-
-/**
- * @brief Set signed 32-bit value
- * @param index Field index
- * @param value Value to set
- *
- * Sets a signed 32-bit field value and marks it as changed.
- */
-void Object::SetInt32Value(uint16 index, int32 value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (m_int32Values[index] != value)
-    {
-        m_int32Values[index] = value;
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set unsigned 32-bit value
- * @param index Field index
- * @param value Value to set
- *
- * Sets an unsigned 32-bit field value and marks it as changed.
- */
-void Object::SetUInt32Value(uint16 index, uint32 value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (m_uint32Values[index] != value)
-    {
-        m_uint32Values[index] = value;
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Update unsigned 32-bit value
- * @param index Field index
- * @param value Value to set
- *
- * Sets an unsigned 32-bit field value and marks it as changed.
- * Does not check if value differs from current.
- */
-void Object::UpdateUInt32Value(uint16 index, uint32 value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    m_uint32Values[index] = value;
-    m_changedValues[index] = true;
-}
-
-/**
- * @brief Set unsigned 64-bit value
- * @param index Field index (uses two consecutive fields)
- * @param value Value to set
- *
- * Sets a 64-bit field value across two consecutive 32-bit fields
- * and marks both as changed.
- */
-void Object::SetUInt64Value(uint16 index, const uint64& value)
-{
-    MANGOS_ASSERT(index + 1 < m_valuesCount || PrintIndexError(index, true));
-    if (*((uint64*) & (m_uint32Values[index])) != value)
-    {
-        m_uint32Values[index] = *((uint32*)&value);
-        m_uint32Values[index + 1] = *(((uint32*)&value) + 1);
-        m_changedValues[index] = true;
-        m_changedValues[index + 1] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set float value
- * @param index Field index
- * @param value Value to set
- *
- * Sets a floating-point field value and marks it as changed.
- */
-void Object::SetFloatValue(uint16 index, float value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (m_floatValues[index] != value)
-    {
-        m_floatValues[index] = value;
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set byte value
- * @param index Field index
- * @param offset Byte offset within the field
- * @param value Value to set
- *
- * Sets a single byte within a 32-bit field and marks it as changed.
- */
-void Object::SetByteValue(uint16 index, uint8 offset, uint8 value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (offset > 4)
-    {
-        sLog.outError("Object::SetByteValue: wrong offset %u", offset);
-        return;
-    }
-
-    if (uint8(m_uint32Values[index] >> (offset * 8)) != value)
-    {
-        m_uint32Values[index] &= ~uint32(uint32(0xFF) << (offset * 8));
-        m_uint32Values[index] |= uint32(uint32(value) << (offset * 8));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set unsigned 16-bit value
- * @param index Field index
- * @param offset 16-bit offset within the field (0 or 1)
- * @param value Value to set
- *
- * Sets a 16-bit value within a 32-bit field and marks it as changed.
- */
-void Object::SetUInt16Value(uint16 index, uint8 offset, uint16 value)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (offset > 2)
-    {
-        sLog.outError("Object::SetUInt16Value: wrong offset %u", offset);
-        return;
-    }
-
-    if (uint16(m_uint32Values[index] >> (offset * 16)) != value)
-    {
-        m_uint32Values[index] &= ~uint32(uint32(0xFFFF) << (offset * 16));
-        m_uint32Values[index] |= uint32(uint32(value) << (offset * 16));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set stat float value
- * @param index Field index
- * @param value Value to set
- *
- * Sets a floating-point stat value, clamping to minimum 0.
- */
-void Object::SetStatFloatValue(uint16 index, float value)
-{
-    if (value < 0)
-    {
-        value = 0.0f;
-    }
-
-    SetFloatValue(index, value);
-}
-
-/**
- * @brief Set stat int32 value
- * @param index Field index
- * @param value Value to set
- *
- * Sets an integer stat value, clamping to minimum 0.
- */
-void Object::SetStatInt32Value(uint16 index, int32 value)
-{
-    if (value < 0)
-    {
-        value = 0;
-    }
-
-    SetUInt32Value(index, uint32(value));
-}
-
-/**
- * @brief Apply modifier to unsigned 32-bit value
- * @param index Field index
- * @param val Modifier value
- * @param apply If true, add modifier; if false, subtract
- *
- * Applies a modifier to a field value, clamping to minimum 0.
- */
-void Object::ApplyModUInt32Value(uint16 index, int32 val, bool apply)
-{
-    int32 cur = GetUInt32Value(index);
-    cur += (apply ? val : -val);
-    if (cur < 0)
-    {
-        cur = 0;
-    }
-    SetUInt32Value(index, cur);
-}
-
-/**
- * @brief Apply modifier to signed 32-bit value
- * @param index Field index
- * @param val Modifier value
- * @param apply If true, add modifier; if false, subtract
- *
- * Applies a modifier to a signed field value.
- */
-void Object::ApplyModInt32Value(uint16 index, int32 val, bool apply)
-{
-    int32 cur = GetInt32Value(index);
-    cur += (apply ? val : -val);
-    SetInt32Value(index, cur);
-}
-
-/**
- * @brief Apply modifier to signed float value
- * @param index Field index
- * @param val Modifier value
- * @param apply If true, add modifier; if false, subtract
- *
- * Applies a modifier to a floating-point field value.
- */
-void Object::ApplyModSignedFloatValue(uint16 index, float  val, bool apply)
-{
-    float cur = GetFloatValue(index);
-    cur += (apply ? val : -val);
-    SetFloatValue(index, cur);
-}
-
-/**
- * @brief Apply modifier to positive float value
- * @param index Field index
- * @param val Modifier value
- * @param apply If true, add modifier; if false, subtract
- *
- * Applies a modifier to a floating-point field value,
- * clamping to minimum 0.
- */
-void Object::ApplyModPositiveFloatValue(uint16 index, float  val, bool apply)
-{
-    float cur = GetFloatValue(index);
-    cur += (apply ? val : -val);
-    if (cur < 0)
-    {
-        cur = 0;
-    }
-    SetFloatValue(index, cur);
-}
-
-/**
- * @brief Set flag in field
- * @param index Field index
- * @param newFlag Flag to set
- *
- * Sets a flag bit in a field using OR operation.
- */
-void Object::SetFlag(uint16 index, uint32 newFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-    uint32 oldval = m_uint32Values[index];
-    uint32 newval = oldval | newFlag;
-
-    if (oldval != newval)
-    {
-        m_uint32Values[index] = newval;
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Remove flag from field
- * @param index Field index
- * @param oldFlag Flag to remove
- *
- * Removes a flag bit from a field using AND with inverted mask.
- */
-void Object::RemoveFlag(uint16 index, uint32 oldFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-    uint32 oldval = m_uint32Values[index];
-    uint32 newval = oldval & ~oldFlag;
-
-    if (oldval != newval)
-    {
-        m_uint32Values[index] = newval;
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set byte flag in field
- * @param index Field index
- * @param offset Byte offset within the field
- * @param newFlag Flag to set
- *
- * Sets a flag bit within a byte of a field.
- */
-void Object::SetByteFlag(uint16 index, uint8 offset, uint8 newFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (offset > 4)
-    {
-        sLog.outError("Object::SetByteFlag: wrong offset %u", offset);
-        return;
-    }
-
-    if (!(uint8(m_uint32Values[index] >> (offset * 8)) & newFlag))
-    {
-        m_uint32Values[index] |= uint32(uint32(newFlag) << (offset * 8));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Remove byte flag from field
- * @param index Field index
- * @param offset Byte offset within the field
- * @param oldFlag Flag to remove
- *
- * Removes a flag bit within a byte of a field.
- */
-void Object::RemoveByteFlag(uint16 index, uint8 offset, uint8 oldFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (offset > 4)
-    {
-        sLog.outError("Object::RemoveByteFlag: wrong offset %u", offset);
-        return;
-    }
-
-    if (uint8(m_uint32Values[index] >> (offset * 8)) & oldFlag)
-    {
-        m_uint32Values[index] &= ~uint32(uint32(oldFlag) << (offset * 8));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Set short flag in field
- * @param index Field index
- * @param highpart If true, use high 16 bits; if false, use low 16 bits
- * @param newFlag Flag to set
- *
- * Sets a flag bit within a 16-bit portion of a field.
- */
-void Object::SetShortFlag(uint16 index, bool highpart, uint16 newFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (!(uint16(m_uint32Values[index] >> (highpart ? 16 : 0)) & newFlag))
-    {
-        m_uint32Values[index] |= uint32(uint32(newFlag) << (highpart ? 16 : 0));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Remove short flag from field
- * @param index Field index
- * @param highpart If true, use high 16 bits; if false, use low 16 bits
- * @param oldFlag Flag to remove
- *
- * Removes a flag bit within a 16-bit portion of a field.
- */
-void Object::RemoveShortFlag(uint16 index, bool highpart, uint16 oldFlag)
-{
-    MANGOS_ASSERT(index < m_valuesCount || PrintIndexError(index, true));
-
-    if (uint16(m_uint32Values[index] >> (highpart ? 16 : 0)) & oldFlag)
-    {
-        m_uint32Values[index] &= ~uint32(uint32(oldFlag) << (highpart ? 16 : 0));
-        m_changedValues[index] = true;
-        MarkForClientUpdate();
-    }
-}
-
-/**
- * @brief Print index error
- * @param index Field index that caused error
- * @param set If true, was a set operation; if false, was a get operation
- * @return Always false
- *
- * Logs an error when attempting to access a nonexistent field.
- */
-bool Object::PrintIndexError(uint32 index, bool set) const
-{
-    sLog.outError("Attempt %s nonexistent value field: %u (count: %u) for object typeid: %u type mask: %u", (set ? "set value to" : "get value from"), index, m_valuesCount, GetTypeId(), m_objectType);
-
-    // ASSERT must fail after function call
-    return false;
-}
-
-/**
- * @brief Print entry error
- * @param descr Description of the invalid operation
- * @return Always false
- *
- * Logs an error when an invalid operation is performed on this object.
- */
-bool Object::PrintEntryError(char const* descr) const
-{
-    sLog.outError("Object Type %u, Entry %u (lowguid %u) with invalid call for %s", GetTypeId(), GetEntry(), GetObjectGuid().GetCounter(), descr);
-
-    // always false for continue assert fail
-    return false;
-}
-
-/**
- * @brief Build update data for player
- * @param pl Target player
- * @param update_players Map of players to their update data
- *
- * Builds update data for the specified player, adding them
- * to the update map if not already present.
- */
-void Object::BuildUpdateDataForPlayer(Player* pl, UpdateDataMapType& update_players)
-{
-    UpdateDataMapType::iterator iter = update_players.find(pl);
-
-    if (iter == update_players.end())
-    {
-        std::pair<UpdateDataMapType::iterator, bool> p = update_players.insert(UpdateDataMapType::value_type(pl, UpdateData(pl->GetMapId())));
-        MANGOS_ASSERT(p.second);
-        iter = p.first;
-    }
-
-    BuildValuesUpdateBlockForPlayer(&iter->second, iter->first);
-}
-
-/**
- * @brief Add to client update list
- *
- * Base implementation logs error and asserts.
- * Derived classes should override this method.
- */
-void Object::AddToClientUpdateList()
-{
-    sLog.outError("Unexpected call of Object::AddToClientUpdateList for object (TypeId: %u Update fields: %u)", GetTypeId(), m_valuesCount);
-    MANGOS_ASSERT(false);
-}
-
-/**
- * @brief Remove from client update list
- *
- * Base implementation logs error and asserts.
- * Derived classes should override this method.
- */
-void Object::RemoveFromClientUpdateList()
-{
-    sLog.outError("Unexpected call of Object::RemoveFromClientUpdateList for object (TypeId: %u Update fields: %u)", GetTypeId(), m_valuesCount);
-    MANGOS_ASSERT(false);
-}
-
-/**
- * @brief Build update data
- * @param update_players Map of players to their update data
- *
- * Base implementation logs error and asserts.
- * Derived classes should override this method.
- */
-void Object::BuildUpdateData(UpdateDataMapType& /*update_players */)
-{
-    sLog.outError("Unexpected call of Object::BuildUpdateData for object (TypeId: %u Update fields: %u)", GetTypeId(), m_valuesCount);
-    MANGOS_ASSERT(false);
-}
-
-/**
- * @brief Mark object for client update
- *
- * Adds the object to the client update list if it's in world
- * and not already marked for update.
- */
-void Object::MarkForClientUpdate()
-{
-    if (m_inWorld)
-    {
-        if (!m_objectUpdated)
-        {
-            AddToClientUpdateList();
-            m_objectUpdated = true;
-        }
-    }
-}
-
-void Object::ForceValuesUpdateAtIndex(uint32 index)
-{
-    m_changedValues[index] = true;
-    if (m_inWorld && !m_objectUpdated)
-    {
-        AddToClientUpdateList();
-        m_objectUpdated = true;
-    }
-}
-
-/**
- * @brief WorldObject constructor
- *
- * Initializes a new WorldObject with default values.
- */
-WorldObject::WorldObject() :
-#ifdef ENABLE_ELUNA
-    elunaEvents(nullptr),
-#endif /* ENABLE_ELUNA */
-    m_transportInfo(NULL),
-    m_currMap(NULL),
-    m_mapId(0), m_InstanceId(0), m_phaseMask(PHASEMASK_NORMAL),
-    m_isActiveObject(false),
-    m_visibilityDistanceOverride(0.0f)
-{
-}
-
-/**
- * @brief WorldObject destructor
- *
- * Cleans up Eluna events if enabled.
- */
-WorldObject::~WorldObject()
-{
-#ifdef ENABLE_ELUNA
-    delete elunaEvents;
-    elunaEvents = nullptr;
-#endif /* ENABLE_ELUNA */
-}
-
-/**
- * @brief Cleanups before delete
- *
- * Removes the object from the world before deletion.
- */
-void WorldObject::CleanupsBeforeDelete()
-{
-    RemoveFromWorld();
-}
-
-/**
- * @brief Update world object
- * @param update_diff Time since last update
- * @param time_diff Time parameter (unused)
- *
- * Updates Eluna events if enabled.
- */
-void WorldObject::Update(uint32 update_diff, uint32 time_diff)
-{
-#ifdef ENABLE_ELUNA
-    if (elunaEvents) // can be null on maps without eluna
-    {
-        elunaEvents->Update(update_diff);
-    }
-#endif /* ENABLE_ELUNA */
-}
-
-/**
- * @brief Create world object
- * @param guidlow Low GUID
- * @param guidhigh High GUID type
- *
- * Creates the world object with the specified GUID.
- */
-void WorldObject::_Create(uint32 guidlow, HighGuid guidhigh, uint32 phaseMask)
-{
-    Object::_Create(guidlow, 0, guidhigh);
-    m_phaseMask = phaseMask;
-}
-
-/**
- * @brief Relocate world object
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param orientation Orientation
- *
- * Moves the object to a new position and orientation.
- * Updates movement info for units.
- */
-void WorldObject::Relocate(float x, float y, float z, float orientation)
-{
-    m_position.x = x;
-    m_position.y = y;
-    m_position.z = z;
-    m_position.o = NormalizeOrientation(orientation);
-
-    if (isType(TYPEMASK_UNIT))
-    {
-        ((Unit*)this)->m_movementInfo.ChangePosition(x, y, z, orientation);
-    }
-}
-
-/**
- * @brief Relocate world object (position only)
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- *
- * Moves the object to a new position without changing orientation.
- * Updates movement info for units.
- */
-void WorldObject::Relocate(float x, float y, float z)
-{
-    m_position.x = x;
-    m_position.y = y;
-    m_position.z = z;
-
-    if (isType(TYPEMASK_UNIT))
-    {
-        ((Unit*)this)->m_movementInfo.ChangePosition(x, y, z, GetOrientation());
-    }
-}
-
-/**
- * @brief Set orientation
- * @param orientation New orientation
- *
- * Sets the object's orientation and updates movement info for units.
- */
-void WorldObject::SetOrientation(float orientation)
-{
-    m_position.o = NormalizeOrientation(orientation);
-
-    if (isType(TYPEMASK_UNIT))
-    {
-        ((Unit*)this)->m_movementInfo.ChangeOrientation(orientation);
-    }
-}
-
-/**
- * @brief Get zone ID
- * @return Zone ID
- *
- * Returns the zone ID based on the object's position.
- */
-uint32 WorldObject::GetZoneId() const
-{
-    return GetTerrain()->GetZoneId(m_position.x, m_position.y, m_position.z);
-}
-
-/**
- * @brief Get area ID
- * @return Area ID
- *
- * Returns the area ID based on the object's position.
- */
-uint32 WorldObject::GetAreaId() const
-{
-    return GetTerrain()->GetAreaId(m_position.x, m_position.y, m_position.z);
-}
-
-/**
- * @brief Get zone and area IDs
- * @param zoneid Output zone ID
- * @param areaid Output area ID
- *
- * Returns both zone and area IDs based on the object's position.
- */
-void WorldObject::GetZoneAndAreaId(uint32& zoneid, uint32& areaid) const
-{
-    GetTerrain()->GetZoneAndAreaId(zoneid, areaid, m_position.x, m_position.y, m_position.z);
-}
-
-/**
- * @brief Get instance data
- * @return Instance data pointer
- *
- * Returns the instance data for the map this object is on.
- */
-InstanceData* WorldObject::GetInstanceData() const
-{
-    return GetMap()->GetInstanceData();
-}
-
-/**
- * @brief Get distance to another object
- * @param obj Target object
- * @return Distance between objects
- *
- * Calculates the 3D distance between this object and another,
- * accounting for bounding radii.
- */
-float WorldObject::GetDistance(const WorldObject* obj) const
-{
-    float dx = GetPositionX() - obj->GetPositionX();
-    float dy = GetPositionY() - obj->GetPositionY();
-    float dz = GetPositionZ() - obj->GetPositionZ();
-    float sizefactor = GetObjectBoundingRadius() + obj->GetObjectBoundingRadius();
-    float dist = sqrt((dx * dx) + (dy * dy) + (dz * dz)) - sizefactor;
-    return (dist > 0 ? dist : 0);
-}
-
-/**
- * @brief Get 2D distance to point
- * @param x X coordinate
- * @param y Y coordinate
- * @return 2D distance to point
- *
- * Calculates the 2D distance between this object and a point,
- * accounting for bounding radius.
- */
-float WorldObject::GetDistance2d(float x, float y) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float sizefactor = GetObjectBoundingRadius();
-    float dist = sqrt((dx * dx) + (dy * dy)) - sizefactor;
-    return (dist > 0 ? dist : 0);
-}
-
-/**
- * @brief Get 3D distance to point
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @return 3D distance to point
- *
- * Calculates the 3D distance between this object and a point,
- * accounting for bounding radius.
- */
-float WorldObject::GetDistance(float x, float y, float z) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float dz = GetPositionZ() - z;
-    float sizefactor = GetObjectBoundingRadius();
-    float dist = sqrt((dx * dx) + (dy * dy) + (dz * dz)) - sizefactor;
-    return (dist > 0 ? dist : 0);
-}
-
-/**
- * @brief Get 2D distance to another object
- * @param obj Target object
- * @return 2D distance to object
- *
- * Calculates the 2D distance between this object and another,
- * accounting for bounding radii.
- */
-float WorldObject::GetDistance2d(const WorldObject* obj) const
-{
-    float dx = GetPositionX() - obj->GetPositionX();
-    float dy = GetPositionY() - obj->GetPositionY();
-    float sizefactor = GetObjectBoundingRadius() + obj->GetObjectBoundingRadius();
-    float dist = sqrt((dx * dx) + (dy * dy)) - sizefactor;
-    return (dist > 0 ? dist : 0);
-}
-
-/**
- * @brief Get vertical distance to another object
- * @param obj Target object
- * @return Vertical distance to object
- *
- * Calculates the vertical (Z-axis) distance between this object
- * and another, accounting for bounding radii.
- */
-float WorldObject::GetDistanceZ(const WorldObject* obj) const
-{
-    float dz = fabs(GetPositionZ() - obj->GetPositionZ());
-    float sizefactor = GetObjectBoundingRadius() + obj->GetObjectBoundingRadius();
-    float dist = dz - sizefactor;
-    return (dist > 0 ? dist : 0);
-}
-
-/**
- * @brief Check if within 3D distance of point
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param dist2compare Distance to compare against
- * @return True if within distance
- *
- * Checks if this object is within the specified 3D distance
- * of the given point.
- */
-bool WorldObject::IsWithinDist3d(float x, float y, float z, float dist2compare) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float dz = GetPositionZ() - z;
-    float distsq = dx * dx + dy * dy + dz * dz;
-
-    float sizefactor = GetObjectBoundingRadius();
-    float maxdist = dist2compare + sizefactor;
-
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Check if within 2D distance of point
- * @param x X coordinate
- * @param y Y coordinate
- * @param dist2compare Distance to compare against
- * @return True if within distance
- *
- * Checks if this object is within the specified 2D distance
- * of the given point.
- */
-bool WorldObject::IsWithinDist2d(float x, float y, float dist2compare) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float distsq = dx * dx + dy * dy;
-
-    float sizefactor = GetObjectBoundingRadius();
-    float maxdist = dist2compare + sizefactor;
-
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Internal check if within distance of object
- * @param obj Target object
- * @param dist2compare Distance to compare against
- * @param is3D If true, check 3D distance; if false, check 2D
- * @return True if within distance
- *
- * Internal helper for distance checking with optional 3D.
- */
-bool WorldObject::_IsWithinDist(WorldObject const* obj, float dist2compare, bool is3D) const
-{
-    float dx = GetPositionX() - obj->GetPositionX();
-    float dy = GetPositionY() - obj->GetPositionY();
-    float distsq = dx * dx + dy * dy;
-    if (is3D)
-    {
-        float dz = GetPositionZ() - obj->GetPositionZ();
-        distsq += dz * dz;
-    }
-    float sizefactor = GetObjectBoundingRadius() + obj->GetObjectBoundingRadius();
-    float maxdist = dist2compare + sizefactor;
-
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Check if within line of sight in map
- * @param obj Target object
- * @return True if within line of sight
- *
- * Checks if this object has line of sight to the target object
- * within the same map.
- */
-namespace
-{
-    // Shift `origin` toward `dest` by min(distance, reach). Mirrors TC's
-    // GetHitSpherePointFor at tc-preservation/src/server/game/Entities/
-    // Object/Object.cpp:1376-1402. Used to start the LoS ray at the
-    // edge of a unit's collision sphere instead of its center so the
-    // ray doesn't immediately self-collide with WMO geometry sitting
-    // right next to the unit (e.g. cathedral columns whose AABB grazes
-    // the unit's footprint). Players have small collision capsules and
-    // are NOT shifted; only non-player units shift by combat reach.
-    void ShiftBySphereRadius(float reach, float destX, float destY, float destZ,
-                             float& x, float& y, float& z)
-    {
-        float dx = destX - x;
-        float dy = destY - y;
-        float dz = destZ - z;
-        float dist = std::sqrt(dx * dx + dy * dy + dz * dz);
-        if (dist < 1e-4f)
-        {
-            return;
-        }
-        float shift = std::min(dist, reach);
-        x += (dx / dist) * shift;
-        y += (dy / dist) * shift;
-        z += (dz / dist) * shift;
-    }
-}
-
-bool WorldObject::IsWithinLOSInMap(const WorldObject* obj, VMAP::ModelIgnoreFlags ignoreFlags) const
-{
-    if (!IsInMap(obj))
-    {
-        return false;
-    }
-
-    // Source: this object's position at head height.
-    float x, y, z;
-    GetPosition(x, y, z);
-    z += 2.0f;
-
-    // Destination: obj's position at head height.
-    float ox, oy, oz;
-    obj->GetPosition(ox, oy, oz);
-    oz += 2.0f;
-
-    // TC's GetHitSpherePointFor pattern: shift non-player endpoints
-    // toward the opposite side by combat reach. Avoids the ray catching
-    // collision geometry sitting at the unit's feet (the Stormwind
-    // Cathedral column-AABB-grazes-player case verified via
-    // .debug losdebug, 2026-05-23).
-    if (GetTypeId() != TYPEID_PLAYER)
-    {
-        if (Unit const* thisUnit = ToUnit())
-        {
-            ShiftBySphereRadius(thisUnit->GetFloatValue(UNIT_FIELD_COMBATREACH),
-                                ox, oy, oz, x, y, z);
-        }
-    }
-    if (obj->GetTypeId() != TYPEID_PLAYER)
-    {
-        if (Unit const* objUnit = obj->ToUnit())
-        {
-            float origX = x, origY = y, origZ = z; // shifted source — fine
-            ShiftBySphereRadius(objUnit->GetFloatValue(UNIT_FIELD_COMBATREACH),
-                                origX, origY, origZ, ox, oy, oz);
-        }
-    }
-
-    return GetMap()->IsInLineOfSight(x, y, z, ox, oy, oz, GetPhaseMask(), ignoreFlags);
-}
-
-/**
- * @brief Check if within line of sight to point
- * @param ox Target X coordinate
- * @param oy Target Y coordinate
- * @param oz Target Z coordinate
- * @param ignoreFlags Caller-supplied filter — e.g. VMAP::ModelIgnoreFlags::M2
- *                    skips MOD_M2 doodads for spell LoS checks.
- * @return True if within line of sight
- *
- * Checks if this object has line of sight to the specified point.
- */
-bool WorldObject::IsWithinLOS(float ox, float oy, float oz, VMAP::ModelIgnoreFlags ignoreFlags) const
-{
-    float x, y, z;
-    GetPosition(x, y, z);
-    z += 2.0f;
-    // Shift non-player source toward the destination point by combat
-    // reach. See ShiftBySphereRadius docs above.
-    if (GetTypeId() != TYPEID_PLAYER)
-    {
-        if (Unit const* thisUnit = ToUnit())
-        {
-            ShiftBySphereRadius(thisUnit->GetFloatValue(UNIT_FIELD_COMBATREACH),
-                                ox, oy, oz + 2.0f, x, y, z);
-        }
-    }
-    return GetMap()->IsInLineOfSight(x, y, z, ox, oy, oz + 2.0f, GetPhaseMask(), ignoreFlags);
-}
-
-/**
- * @brief Compare distance order to two objects
- * @param obj1 First object
- * @param obj2 Second object
- * @param is3D If true, use 3D distance; if false, use 2D
- * @return True if obj1 is closer than obj2
- *
- * Compares distances to two objects to determine which is closer.
- */
-bool WorldObject::GetDistanceOrder(WorldObject const* obj1, WorldObject const* obj2, bool is3D /* = true */) const
-{
-    float dx1 = GetPositionX() - obj1->GetPositionX();
-    float dy1 = GetPositionY() - obj1->GetPositionY();
-    float distsq1 = dx1 * dx1 + dy1 * dy1;
-    if (is3D)
-    {
-        float dz1 = GetPositionZ() - obj1->GetPositionZ();
-        distsq1 += dz1 * dz1;
-    }
-
-    float dx2 = GetPositionX() - obj2->GetPositionX();
-    float dy2 = GetPositionY() - obj2->GetPositionY();
-    float distsq2 = dx2 * dx2 + dy2 * dy2;
-    if (is3D)
-    {
-        float dz2 = GetPositionZ() - obj2->GetPositionZ();
-        distsq2 += dz2 * dz2;
-    }
-
-    return distsq1 < distsq2;
-}
-
-/**
- * @brief Check if within range of object
- * @param obj Target object
- * @param minRange Minimum distance
- * @param maxRange Maximum distance
- * @param is3D If true, use 3D distance; if false, use 2D
- * @return True if within range
- *
- * Checks if this object is within the specified distance range
- * of the target object.
- */
-bool WorldObject::IsInRange(WorldObject const* obj, float minRange, float maxRange, bool is3D /* = true */) const
-{
-    float dx = GetPositionX() - obj->GetPositionX();
-    float dy = GetPositionY() - obj->GetPositionY();
-    float distsq = dx * dx + dy * dy;
-    if (is3D)
-    {
-        float dz = GetPositionZ() - obj->GetPositionZ();
-        distsq += dz * dz;
-    }
-
-    float sizefactor = GetObjectBoundingRadius() + obj->GetObjectBoundingRadius();
-
-    // check only for real range
-    if (minRange > 0.0f)
-    {
-        float mindist = minRange + sizefactor;
-        if (distsq < mindist * mindist)
-        {
-            return false;
-        }
-    }
-
-    float maxdist = maxRange + sizefactor;
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Check if within 2D range of point
- * @param x X coordinate
- * @param y Y coordinate
- * @param minRange Minimum distance
- * @param maxRange Maximum distance
- * @return True if within range
- *
- * Checks if this object is within the specified 2D distance range
- * of the target point.
- */
-bool WorldObject::IsInRange2d(float x, float y, float minRange, float maxRange) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float distsq = dx * dx + dy * dy;
-
-    float sizefactor = GetObjectBoundingRadius();
-
-    // check only for real range
-    if (minRange > 0.0f)
-    {
-        float mindist = minRange + sizefactor;
-        if (distsq < mindist * mindist)
-        {
-            return false;
-        }
-    }
-
-    float maxdist = maxRange + sizefactor;
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Check if within 3D range of point
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param minRange Minimum distance
- * @param maxRange Maximum distance
- * @return True if within range
- *
- * Checks if this object is within the specified 3D distance range
- * of the target point.
- */
-bool WorldObject::IsInRange3d(float x, float y, float z, float minRange, float maxRange) const
-{
-    float dx = GetPositionX() - x;
-    float dy = GetPositionY() - y;
-    float dz = GetPositionZ() - z;
-    float distsq = dx * dx + dy * dy + dz * dz;
-
-    float sizefactor = GetObjectBoundingRadius();
-
-    // check only for real range
-    if (minRange > 0.0f)
-    {
-        float mindist = minRange + sizefactor;
-        if (distsq < mindist * mindist)
-        {
-            return false;
-        }
-    }
-
-    float maxdist = maxRange + sizefactor;
-    return distsq < maxdist * maxdist;
-}
-
-/**
- * @brief Get angle to object
- * @param obj Target object
- * @return Angle in radians (0 to 2*PI)
- *
- * Returns the angle from this object to the target object.
- */
-float WorldObject::GetAngle(const WorldObject* obj) const
-{
-    if (!obj)
-    {
-        return 0.0f;
-    }
-
-    // Rework the assert, when more cases where such a call can happen have been fixed
-    // MANGOS_ASSERT(obj != this || PrintEntryError("GetAngle (for self)"));
-    if (obj == this)
-    {
-        sLog.outError("INVALID CALL for GetAngle for %s", obj->GetGuidStr().c_str());
-        return 0.0f;
-    }
-    return GetAngle(obj->GetPositionX(), obj->GetPositionY());
-}
-
-/**
- * @brief Get angle to point
- * @param x X coordinate
- * @param y Y coordinate
- * @return Angle in radians (0 to 2*PI)
- *
- * Returns the angle from this object to the specified point.
- */
-float WorldObject::GetAngle(const float x, const float y) const
-{
-    float dx = x - GetPositionX();
-    float dy = y - GetPositionY();
-
-    float ang = atan2(dy, dx);                              // returns value between -Pi..Pi
-    ang = (ang >= 0) ? ang : 2 * M_PI_F + ang;
-    return ang;
-}
-
-/**
- * @brief Check if object is within arc
- * @param arcangle Arc angle in radians
- * @param obj Target object
- * @return True if object is within arc
- *
- * Checks if the target object is within the specified arc
- * in front of this object.
- */
-bool WorldObject::HasInArc(const float arcangle, const WorldObject* obj) const
-{
-    // always have self in arc
-    if (obj == this)
-    {
-        return true;
-    }
-
-    float arc = arcangle;
-
-    // move arc to range 0.. 2*pi
-    arc = NormalizeOrientation(arc);
-
-    float angle = GetAngle(obj);
-    angle -= m_position.o;
-
-    // move angle to range -pi ... +pi
-    angle = NormalizeOrientation(angle);
-    if (angle > M_PI_F)
-    {
-        angle -= 2.0f * M_PI_F;
-    }
-
-    float lborder =  -1 * (arc / 2.0f);                     // in range -pi..0
-    float rborder = (arc / 2.0f);                           // in range 0..pi
-    return ((angle >= lborder) && (angle <= rborder));
-}
-
-/**
- * @brief Check if target is in front in same map
- * @param target Target object
- * @param distance Maximum distance
- * @param arc Arc angle in radians
- * @return True if target is in front
- *
- * Checks if the target is in front of this object within
- * the specified distance and arc, in the same map.
- */
-bool WorldObject::IsInFrontInMap(WorldObject const* target, float distance,  float arc) const
-{
-    return IsWithinDistInMap(target, distance) && HasInArc(arc, target);
-}
-
-/**
- * @brief Check if target is in back in same map
- * @param target Target object
- * @param distance Maximum distance
- * @param arc Arc angle in radians
- * @return True if target is in back
- *
- * Checks if the target is behind this object within
- * the specified distance and arc, in the same map.
- */
-bool WorldObject::IsInBackInMap(WorldObject const* target, float distance, float arc) const
-{
-    return IsWithinDistInMap(target, distance) && !HasInArc(2 * M_PI_F - arc, target);
-}
-
-/**
- * @brief Check if target is in front
- * @param target Target object
- * @param distance Maximum distance
- * @param arc Arc angle in radians
- * @return True if target is in front
- *
- * Checks if the target is in front of this object within
- * the specified distance and arc.
- */
-bool WorldObject::IsInFront(WorldObject const* target, float distance,  float arc) const
-{
-    return IsWithinDist(target, distance) && HasInArc(arc, target);
-}
-
-/**
- * @brief Check if target is in back
- * @param target Target object
- * @param distance Maximum distance
- * @param arc Arc angle in radians
- * @return True if target is in back
- *
- * Checks if the target is behind this object within
- * the specified distance and arc.
- */
-bool WorldObject::IsInBack(WorldObject const* target, float distance, float arc) const
-{
-    return IsWithinDist(target, distance) && !HasInArc(2 * M_PI_F - arc, target);
-}
-
-/**
- * @brief Get random point near position
- * @param x Center X coordinate
- * @param y Center Y coordinate
- * @param z Center Z coordinate
- * @param distance Maximum distance from center
- * @param rand_x Output random X coordinate
- * @param rand_y Output random Y coordinate
- * @param rand_z Output random Z coordinate
- * @param minDist Minimum distance from center
- * @param ori Optional orientation to use instead of random
- *
- * Generates a random point within the specified distance
- * of the center position.
- */
-void WorldObject::GetRandomPoint(float x, float y, float z, float distance, float& rand_x, float& rand_y, float& rand_z, float minDist /*=0.0f*/, float const* ori /*=NULL*/) const
-{
-    if (distance == 0)
-    {
-        rand_x = x;
-        rand_y = y;
-        rand_z = z;
-        return;
-    }
-
-    // angle to face `obj` to `this`
-    float angle;
-    if (!ori)
-    {
-        angle = rand_norm_f() * 2 * M_PI_F;
-    }
-    else
-    {
-        angle = *ori;
-    }
-
-    float new_dist;
-    if (minDist == 0.0f)
-    {
-        new_dist = rand_norm_f() * distance;
-    }
-    else
-    {
-        new_dist = minDist + rand_norm_f() * (distance - minDist);
-    }
-
-    rand_x = x + new_dist * cos(angle);
-    rand_y = y + new_dist * sin(angle);
-    rand_z = z;
-
-    MaNGOS::NormalizeMapCoord(rand_x);
-    MaNGOS::NormalizeMapCoord(rand_y);
-    UpdateGroundPositionZ(rand_x, rand_y, rand_z);          // update to LOS height if available
-}
-
-/**
- * @brief Update ground position Z
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z-coordinate to update
- *
- * Updates the Z-coordinate to the ground height at the
- * specified position.
- */
-void WorldObject::UpdateGroundPositionZ(float x, float y, float& z) const
-{
-    float new_z = GetMap()->GetHeight(GetPhaseMask(), x, y, z);
-    if (new_z > INVALID_HEIGHT)
-    {
-        z = new_z + 0.05f;                                   // just to be sure that we are not a few pixel under the surface
-    }
-}
-
-/**
- * @brief Update allowed position Z
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z-coordinate to update
- * @param atMap Map to use for height calculation (optional)
- *
- * Updates the Z-coordinate to a valid height based on the
- * object's movement capabilities (flying, swimming, etc.).
- */
-void WorldObject::UpdateAllowedPositionZ(float x, float y, float& z, Map* atMap /*=NULL*/) const
-{
-    if (!atMap)
-    {
-        atMap = GetMap();
-    }
-
-    switch (GetTypeId())
-    {
-        case TYPEID_UNIT:
-        {
-            // non fly unit don't must be in air
-            // non swim unit must be at ground (mostly speedup, because it don't must be in water and water level check less fast
-            if (!((Creature const*)this)->CanFly())
-            {
-                bool canSwim = ((Creature const*)this)->CanSwim();
-                float ground_z = z;
-                float max_z = canSwim
-                              ? atMap->GetTerrain()->GetWaterOrGroundLevel(x, y, z, &ground_z, !((Unit const*)this)->HasAuraType(SPELL_AURA_WATER_WALK))
-                              : ((ground_z = atMap->GetHeight(GetPhaseMask(), x, y, z)));
-                if (max_z > INVALID_HEIGHT)
-                {
-                    if (z > max_z)
-                    {
-                        z = max_z;
-                    }
-                    else if (z < ground_z)
-                    {
-                        z = ground_z;
-                    }
-                }
-            }
-            else
-            {
-                float ground_z = atMap->GetHeight(GetPhaseMask(), x, y, z);
-                if (z < ground_z)
-                {
-                    z = ground_z;
-                }
-            }
-            break;
-        }
-        case TYPEID_PLAYER:
-        {
-            // for server controlled moves player work same as creature (but it can always swim)
-            if (!((Player const*)this)->CanFly())
-            {
-                float ground_z = z;
-                float max_z = atMap->GetTerrain()->GetWaterOrGroundLevel(x, y, z, &ground_z, !((Unit const*)this)->HasAuraType(SPELL_AURA_WATER_WALK));
-                if (max_z > INVALID_HEIGHT)
-                {
-                    if (z > max_z)
-                    {
-                        z = max_z;
-                    }
-                    else if (z < ground_z)
-                    {
-                        z = ground_z;
-                    }
-                }
-            }
-            else
-            {
-                float ground_z = atMap->GetHeight(GetPhaseMask(), x, y, z);
-                if (z < ground_z)
-                {
-                    z = ground_z;
-                }
-            }
-            break;
-        }
-        default:
-        {
-            float ground_z = atMap->GetHeight(GetPhaseMask(), x, y, z);
-            if (ground_z > INVALID_HEIGHT)
-            {
-                z = ground_z;
-            }
-            break;
-        }
-    }
-}
-
-/**
- * @brief Check if position is valid
- * @return True if position is valid
- *
- * Checks if the object's current position is valid.
- */
-bool WorldObject::IsPositionValid() const
-{
-    return MaNGOS::IsValidMapCoord(m_position.x, m_position.y, m_position.z, m_position.o);
-}
-
-/**
- * @brief Send monster say message
- * @param text Text to say
- * @param language Language (unused)
- * @param target Target unit
- *
- * Sends a monster say message to nearby players.
- */
-void WorldObject::MonsterSay(const char* text, uint32 language, Unit const* target) const
-{
-    WorldPacket data(SMSG_MESSAGECHAT, 200);
-    ChatHandler::BuildChatPacket(data, CHAT_MSG_MONSTER_SAY, text, LANG_UNIVERSAL, CHAT_TAG_NONE, GetObjectGuid(), GetName(),
-                                 target ? target->GetObjectGuid() : ObjectGuid(), target ? target->GetName() : "");
-    SendMessageToSetInRange(&data, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_SAY), true);
-}
-
-/**
- * @brief Send monster yell message
- * @param text Text to yell
- * @param language Language (unused)
- * @param target Target unit
- *
- * Sends a monster yell message to nearby players.
- */
-void WorldObject::MonsterYell(const char* text, uint32 language, Unit const* target) const
-{
-    WorldPacket data(SMSG_MESSAGECHAT, 200);
-    ChatHandler::BuildChatPacket(data, CHAT_MSG_MONSTER_YELL, text, LANG_UNIVERSAL, CHAT_TAG_NONE, GetObjectGuid(), GetName(),
-                                 target ? target->GetObjectGuid() : ObjectGuid(), target ? target->GetName() : "");
-    SendMessageToSetInRange(&data, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_YELL), true);
-}
-
-/**
- * @brief Send monster text emote message
- * @param text Text to emote
- * @param target Target unit
- * @param IsBossEmote If true, use boss emote range
- *
- * Sends a monster text emote message to nearby players.
- */
-void WorldObject::MonsterTextEmote(const char* text, Unit const* target, bool IsBossEmote) const
-{
-    WorldPacket data(SMSG_MESSAGECHAT, 200);
-    ChatHandler::BuildChatPacket(data, IsBossEmote ? CHAT_MSG_RAID_BOSS_EMOTE : CHAT_MSG_MONSTER_EMOTE, text, LANG_UNIVERSAL, CHAT_TAG_NONE, GetObjectGuid(), GetName(),
-                                 target ? target->GetObjectGuid() : ObjectGuid(), target ? target->GetName() : "");
-    SendMessageToSetInRange(&data, sWorld.getConfig(IsBossEmote ? CONFIG_FLOAT_LISTEN_RANGE_YELL : CONFIG_FLOAT_LISTEN_RANGE_TEXTEMOTE), true);
-}
-
-/**
- * @brief Send monster whisper message
- * @param text Text to whisper
- * @param target Target unit
- * @param IsBossWhisper If true, use boss whisper
- *
- * Sends a monster whisper message to the target player.
- */
-void WorldObject::MonsterWhisper(const char* text, Unit const* target, bool IsBossWhisper) const
-{
-    if (!target || target->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    WorldPacket data(SMSG_MESSAGECHAT, 200);
-    ChatHandler::BuildChatPacket(data, IsBossWhisper ? CHAT_MSG_RAID_BOSS_WHISPER : CHAT_MSG_MONSTER_WHISPER, text, LANG_UNIVERSAL, CHAT_TAG_NONE, GetObjectGuid(), GetName(),
-                                 target->GetObjectGuid(), target->GetName());
-    ((Player*)target)->GetSession()->SendPacket(&data);
-}
-
-namespace MaNGOS
-{
-    /**
-     * @brief Monster chat builder functor
-     *
-     * Builds localized chat packets for monster messages.
-     */
-    class MonsterChatBuilder
-    {
-        public:
-            /**
-             * @brief Constructor
-             * @param obj Source object
-             * @param msgtype Chat message type
-             * @param textData Localized text data
-             * @param language Language
-             * @param target Target unit
-             */
-            MonsterChatBuilder(WorldObject const& obj, ChatMsg msgtype, MangosStringLocale const* textData, Language language, Unit const* target)
-                : i_object(obj), i_msgtype(msgtype), i_textData(textData), i_language(language), i_target(target) {}
-
-            /**
-             * @brief Build chat packet
-             * @param data Packet to build
-             * @param loc_idx Locale index
-             */
-            void operator()(WorldPacket& data, int32 loc_idx)
-            {
-                char const* text = NULL;
-                if ((int32)i_textData->Content.size() > loc_idx + 1 && !i_textData->Content[loc_idx + 1].empty())
-                {
-                    text = i_textData->Content[loc_idx + 1].c_str();
-                }
-                else
-                {
-                    text = i_textData->Content[0].c_str();
-                }
-
-                ChatHandler::BuildChatPacket(data, i_msgtype, text, i_language, CHAT_TAG_NONE, i_object.GetObjectGuid(), i_object.GetNameForLocaleIdx(loc_idx),
-                                             i_target ? i_target->GetObjectGuid() : ObjectGuid(), i_target ? i_target->GetNameForLocaleIdx(loc_idx) : "");
-            }
-
-        private:
-            WorldObject const& i_object; ///< Source object
-            ChatMsg i_msgtype; ///< Chat message type
-            MangosStringLocale const* i_textData; ///< Localized text data
-            Language i_language; ///< Language
-            Unit const* i_target; ///< Target unit
-    };
-}                                                           // namespace MaNGOS
-
-/**
- * @brief Send localized text around source
- * @param source Source object
- * @param textData Localized text data
- * @param msgtype Chat message type
- * @param language Language
- * @param target Target unit
- * @param range Range to send message
- *
- * Helper function to create localized chat around a source.
- */
-/**
- * @brief Sends localized monster chat packets to players around a source object.
- *
- * @param source The source world object.
- * @param textData The localized text definition.
- * @param msgtype The chat packet type.
- * @param language The chat language.
- * @param target Optional target unit.
- * @param range Broadcast range.
- */
-void _DoLocalizedTextAround(WorldObject const* source, MangosStringLocale const* textData, ChatMsg msgtype, Language language, Unit const* target, float range)
-{
-    MaNGOS::MonsterChatBuilder say_build(*source, msgtype, textData, language, target);
-    MaNGOS::LocalizedPacketDo<MaNGOS::MonsterChatBuilder> say_do(say_build);
-    MaNGOS::CameraDistWorker<MaNGOS::LocalizedPacketDo<MaNGOS::MonsterChatBuilder> > say_worker(source, range, say_do);
-    Cell::VisitWorldObjects(source, say_worker, range);
-}
-
-/**
- * @brief Send monster text
- * @param textData Localized text data
- * @param target Target unit
- *
- * Sends a text message associated with a MangosString,
- * localized for each player's locale.
- */
-/**
- * @brief Sends a localized monster text defined by a Mangos string entry.
- *
- * @param textData The localized text definition.
- * @param target Optional target unit.
- */
-void WorldObject::MonsterText(MangosStringLocale const* textData, Unit const* target) const
-{
-    MANGOS_ASSERT(textData);
-
-    switch (textData->Type)
-    {
-        case CHAT_TYPE_SAY:
-            _DoLocalizedTextAround(this, textData, CHAT_MSG_MONSTER_SAY, textData->LanguageId, target, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_SAY));
-            break;
-        case CHAT_TYPE_YELL:
-            _DoLocalizedTextAround(this, textData, CHAT_MSG_MONSTER_YELL, textData->LanguageId, target, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_YELL));
-            break;
-        case CHAT_TYPE_TEXT_EMOTE:
-            _DoLocalizedTextAround(this, textData, CHAT_MSG_MONSTER_EMOTE, LANG_UNIVERSAL, target, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_TEXTEMOTE));
-            break;
-        case CHAT_TYPE_BOSS_EMOTE:
-            _DoLocalizedTextAround(this, textData, CHAT_MSG_RAID_BOSS_EMOTE, LANG_UNIVERSAL, target, sWorld.getConfig(CONFIG_FLOAT_LISTEN_RANGE_YELL));
-            break;
-        case CHAT_TYPE_WHISPER:
-        {
-            if (!target || target->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-            MaNGOS::MonsterChatBuilder say_build(*this, CHAT_MSG_MONSTER_WHISPER, textData, LANG_UNIVERSAL, target);
-            MaNGOS::LocalizedPacketDo<MaNGOS::MonsterChatBuilder> say_do(say_build);
-            say_do((Player*)target);
-            break;
-        }
-        case CHAT_TYPE_BOSS_WHISPER:
-        {
-            if (!target || target->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-            MaNGOS::MonsterChatBuilder say_build(*this, CHAT_MSG_RAID_BOSS_WHISPER, textData, LANG_UNIVERSAL, target);
-            MaNGOS::LocalizedPacketDo<MaNGOS::MonsterChatBuilder> say_do(say_build);
-            say_do((Player*)target);
-            break;
-        }
-        case CHAT_TYPE_ZONE_YELL:
-        {
-            MaNGOS::MonsterChatBuilder say_build(*this, CHAT_MSG_MONSTER_YELL, textData, textData->LanguageId, target);
-            MaNGOS::LocalizedPacketDo<MaNGOS::MonsterChatBuilder> say_do(say_build);
-            uint32 zoneid = GetZoneId();
-            Map::PlayerList const& pList = GetMap()->GetPlayers();
-            for (Map::PlayerList::const_iterator itr = pList.begin(); itr != pList.end(); ++itr)
-                if (itr->getSource()->GetZoneId() == zoneid)
-                {
-                    say_do(itr->getSource());
-                }
-            break;
-        }
-    }
-}
-
-/**
- * @brief Broadcasts a packet to all players in the object's visibility set.
- *
- * @param data The packet to send.
- * @param bToSelf Unused self-delivery flag.
- */
-void WorldObject::SendMessageToSet(WorldPacket* data, bool /*bToSelf*/) const
-{
-    // if object is in world, map for it already created!
-    if (IsInWorld())
-    {
-        GetMap()->MessageBroadcast(this, data);
-    }
-}
-
-/**
- * @brief Broadcasts a packet to players within a specified range.
- *
- * @param data The packet to send.
- * @param dist The broadcast distance.
- * @param bToSelf Unused self-delivery flag.
- */
-void WorldObject::SendMessageToSetInRange(WorldPacket* data, float dist, bool /*bToSelf*/) const
-{
-    // if object is in world, map for it already created!
-    if (IsInWorld())
-    {
-        GetMap()->MessageDistBroadcast(this, data, dist);
-    }
-}
-
-/**
- * @brief Broadcasts a packet to visible players except one receiver.
- *
- * @param data The packet to send.
- * @param skipped_receiver The player to exclude.
- */
-void WorldObject::SendMessageToSetExcept(WorldPacket* data, Player const* skipped_receiver) const
-{
-    // if object is in world, map for it already created!
-    if (IsInWorld())
-    {
-        MaNGOS::MessageDelivererExcept notifier(this, data, skipped_receiver);
-        Cell::VisitWorldObjects(this, notifier, GetMap()->GetBroadcastRadius());
-    }
-}
-
-void WorldObject::SendObjectDeSpawnAnim(ObjectGuid guid)
-{
-    WorldPacket data(SMSG_GAMEOBJECT_DESPAWN_ANIM, 8);
-    data << ObjectGuid(guid);
-    SendMessageToSet(&data, true);
-}
-
-void WorldObject::SendGameObjectCustomAnim(ObjectGuid guid, uint32 animId /*= 0*/)
-{
-    WorldPacket data(SMSG_GAMEOBJECT_CUSTOM_ANIM, 8 + 4);
-    data << ObjectGuid(guid);
-    data << uint32(animId);
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Assigns the current map context to the world object.
- *
- * @param map The map to assign.
- */
-void WorldObject::SetMap(Map* map)
-{
-    MANGOS_ASSERT(map);
-    m_currMap = map;
-    // lets save current map's Id/instanceId
-    m_mapId = map->GetId();
-    m_InstanceId = map->GetInstanceId();
-}
-
-/**
- * @brief Reset map
- *
- * Resets the map reference for this object.
- */
-/**
- * @brief Resets the world object's map state.
- */
-void WorldObject::ResetMap()
-{
-    m_currMap = NULL;
-}
-
-TerrainInfo const* WorldObject::GetTerrain() const
-{
-    MANGOS_ASSERT(m_currMap);
-    return m_currMap->GetTerrain();
-}
-
-/**
- * @brief Add object to remove list
- *
- * Adds this object to the map's remove list for cleanup.
- */
-/**
- * @brief Schedules the object for removal from the map.
- */
-void WorldObject::AddObjectToRemoveList()
-{
-    GetMap()->AddObjectToRemoveList(this);
-}
-
-/**
- * @brief Summon creature
- * @param id Creature entry ID
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param ang Orientation
- * @param spwtype Temporary spawn type
- * @param despwtime Despawn time
- * @param asActiveObject If true, set as active object
- * @param setRun If true, set run mode
- * @return Summoned creature pointer or NULL
- *
- * Summons a creature at the specified position.
- */
-/**
- * @brief Summons a temporary creature near or at the requested position.
- *
- * @param id The creature template id.
- * @param x The summon x coordinate.
- * @param y The summon y coordinate.
- * @param z The summon z coordinate.
- * @param ang The summon orientation.
- * @param spwtype The temporary spawn type.
- * @param despwtime The despawn time in milliseconds.
- * @param asActiveObject true to mark the summon as active.
- * @param setRun true to make the summon run.
- * @return The summoned creature, or null on failure.
- */
-Creature* WorldObject::SummonCreature(uint32 id, float x, float y, float z, float ang, TempSpawnType spwtype, uint32 despwtime, bool asActiveObject, bool setRun)
-{
-    CreatureInfo const* cinfo = ObjectMgr::GetCreatureTemplate(id);
-    if (!cinfo)
-    {
-        sLog.outErrorDb("WorldObject::SummonCreature: Creature (Entry: %u) not existed for summoner: %s. ", id, GetGuidStr().c_str());
-        return NULL;
-    }
-
-    TemporarySummon* pCreature = new TemporarySummon(GetObjectGuid());
-
-    Team team = TEAM_NONE;
-    if (GetTypeId() == TYPEID_PLAYER)
-    {
-        team = ((Player*)this)->GetTeam();
-    }
-
-    CreatureCreatePos pos(GetMap(), x, y, z, ang, GetPhaseMask());
-
-    if (x == 0.0f && y == 0.0f && z == 0.0f)
-    {
-        pos = CreatureCreatePos(this, GetOrientation(), CONTACT_DISTANCE, ang);
-    }
-
-    if (!pCreature->Create(GetMap()->GenerateLocalLowGuid(cinfo->GetHighGuid()), pos, cinfo, team))
-    {
-        delete pCreature;
-        return NULL;
-    }
-
-    pCreature->SetRespawnCoord(pos);
-
-    // Set run or walk before any other movement starts
-    pCreature->SetWalk(!setRun);
-
-    // Active state set before added to map
-    pCreature->SetActiveObjectState(asActiveObject);
-
-    pCreature->Summon(spwtype, despwtime);                  // Also initializes the AI and MMGen
-
-    if (GetTypeId() == TYPEID_UNIT && ((Creature*)this)->AI())
-    {
-        ((Creature*)this)->AI()->JustSummoned(pCreature);
-    }
-
-#ifdef ENABLE_ELUNA
-    if (Unit* summoner = ToUnit())
-    {
-        if (Eluna* e = GetEluna())
-        {
-            e->OnSummoned(pCreature, summoner);
-        }
-    }
-#endif /* ENABLE_ELUNA */
-
-    // Creature Linking, Initial load is handled like respawn
-    if (pCreature->IsLinkingEventTrigger())
-    {
-        GetMap()->GetCreatureLinkingHolder()->DoCreatureLinkingEvent(LINKING_EVENT_RESPAWN, pCreature);
-    }
-
-    // return the creature therewith the summoner has access to it
-    return pCreature;
-}
-
-/**
- * @brief Summon game object
- * @param id Game object entry ID
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param angle Orientation
- * @param despwtime Despawn time in milliseconds
- * @return Summoned game object pointer or NULL
- *
- * Summons a game object at the specified position.
- */
-/**
- * @brief Summons a temporary game object at the requested position.
- *
- * @param id The gameobject entry id.
- * @param x The summon x coordinate.
- * @param y The summon y coordinate.
- * @param z The summon z coordinate.
- * @param angle The summon orientation.
- * @param despwtime The despawn time in milliseconds.
- * @return The summoned game object, or null on failure.
- */
-GameObject* WorldObject::SummonGameObject(uint32 id, float x, float y, float z, float angle, uint32 despwtime)
-{
-    GameObject* pGameObj = new GameObject;
-
-    Map *map = GetMap();
-
-    if (!map)
-    {
-        return NULL;
-    }
-
-    if (!pGameObj->Create(map->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), id, map, GetPhaseMask(), x, y, z, angle))
-    {
-        delete pGameObj;
-        return NULL;
-    }
-
-    pGameObj->SetRespawnTime(despwtime/IN_MILLISECONDS);
-
-    map->Add(pGameObj);
-    pGameObj->AIM_Initialize();
-
-    return pGameObj;
-}
-
-// how much space should be left in front of/ behind a mob that already uses a space
-#define OCCUPY_POS_DEPTH_FACTOR                          1.8f
-
-namespace MaNGOS
-{
-    /**
-     * @brief Near used position functor
-     *
-     * Checks for used positions near an object for position selection.
-     */
-    class NearUsedPosDo
-    {
-        public:
-            /**
-             * @brief Constructor
-             * @param obj Source object
-             * @param searcher Object searching for position
-             * @param absAngle Absolute angle
-             * @param selector Position selector
-             */
-            NearUsedPosDo(WorldObject const& obj, WorldObject const* searcher, float absAngle, ObjectPosSelector& selector)
-                : i_object(obj), i_searcher(searcher), i_absAngle(NormalizeOrientation(absAngle)), i_selector(selector) {}
-
-            void operator()(Corpse*) const {}
-            void operator()(DynamicObject*) const {}
-
-            /**
-             * @brief Process creature
-             * @param c Creature to process
-             */
-            void operator()(Creature* c) const
-            {
-                // skip self or target
-                if (c == i_searcher || c == &i_object)
-                {
-                    return;
-                }
-
-                float x, y, z;
-
-                if (c->IsStopped() || !c->GetMotionMaster()->GetDestination(x, y, z))
-                {
-                    x = c->GetPositionX();
-                    y = c->GetPositionY();
-                }
-
-                add(c, x, y);
-            }
-
-            /**
-             * @brief Process generic unit
-             * @param u Unit to process
-             */
-            template<class T>
-            void operator()(T* u) const
-            {
-                // skip self or target
-                if (u == i_searcher || u == &i_object)
-                {
-                    return;
-                }
-
-                float x, y;
-
-                x = u->GetPositionX();
-                y = u->GetPositionY();
-
-                add(u, x, y);
-            }
-
-            /**
-             * @brief Add used position
-             * @param u Object to add
-             * @param x X coordinate
-             * @param y Y coordinate
-             *
-             * Adds a used position to the selector.
-             */
-            void add(WorldObject* u, float x, float y) const
-            {
-                float dx = i_object.GetPositionX() - x;
-                float dy = i_object.GetPositionY() - y;
-                float dist2d = sqrt((dx * dx) + (dy * dy));
-
-                // It is ok for the objects to require a bit more space
-                float delta = u->GetObjectBoundingRadius();
-                if (i_selector.m_searchPosFor && i_selector.m_searchPosFor != u)
-                {
-                    delta += i_selector.m_searchPosFor->GetObjectBoundingRadius();
-                }
-
-                delta *= OCCUPY_POS_DEPTH_FACTOR;           // Increase by factor
-
-                // u is too near/far away from i_object. Do not consider it to occupy space
-                if (fabs(i_selector.m_searcherDist - dist2d) > delta)
-                {
-                    return;
-                }
-
-                float angle = i_object.GetAngle(u) - i_absAngle;
-
-                // move angle to range -pi ... +pi, range before is -2Pi..2Pi
-                if (angle > M_PI_F)
-                {
-                    angle -= 2.0f * M_PI_F;
-                }
-                else if (angle < -M_PI_F)
-                {
-                    angle += 2.0f * M_PI_F;
-                }
-
-                i_selector.AddUsedArea(u, angle, dist2d);
-            }
-        private:
-            WorldObject const& i_object;
-            WorldObject const* i_searcher;
-            float              i_absAngle;
-            ObjectPosSelector& i_selector;
-    };
-}
-
-/**
- * @brief Get 2D point near object
- * @param x Output X coordinate
- * @param y Output Y coordinate
- * @param distance2d Distance from object
- * @param absAngle Absolute angle
- *
- * Calculates a 2D point at the specified distance and angle
- * from this object.
- */
-/**
- * @brief Computes a 2D point at a given distance and angle from the object.
- *
- * @param x Receives the resulting x coordinate.
- * @param y Receives the resulting y coordinate.
- * @param distance2d The radial distance.
- * @param absAngle The absolute angle.
- */
-void WorldObject::GetNearPoint2D(float& x, float& y, float distance2d, float absAngle) const
-{
-    x = GetPositionX() + distance2d * cos(absAngle);
-    y = GetPositionY() + distance2d * sin(absAngle);
-
-    MaNGOS::NormalizeMapCoord(x);
-    MaNGOS::NormalizeMapCoord(y);
-}
-
-/**
- * @brief Finds a nearby point while accounting for collisions and line of sight.
- *
- * @param searcher The object requesting the position.
- * @param x Receives the resulting x coordinate.
- * @param y Receives the resulting y coordinate.
- * @param z Receives the resulting z coordinate.
- * @param searcher_bounding_radius The requester's bounding radius.
- * @param distance2d The desired distance from this object.
- * @param absAngle The preferred absolute angle.
- */
-void WorldObject::GetNearPoint(WorldObject const* searcher, float& x, float& y, float& z, float searcher_bounding_radius, float distance2d, float absAngle) const
-{
-    GetNearPoint2D(x, y, distance2d, absAngle);
-    const float init_z = z = GetPositionZ();
-
-    // if detection disabled, return first point
-    if (!sWorld.getConfig(CONFIG_BOOL_DETECT_POS_COLLISION))
-    {
-        if (searcher)
-        {
-            searcher->UpdateAllowedPositionZ(x, y, z, GetMap());       // update to LOS height if available
-        }
-        else
-        {
-            UpdateGroundPositionZ(x, y, z);
-        }
-        return;
-    }
-
-    // or remember first point
-    float first_x = x;
-    float first_y = y;
-    bool first_los_conflict = false;                        // first point LOS problems
-
-    const float dist = distance2d + searcher_bounding_radius + GetObjectBoundingRadius();
-
-    // prepare selector for work
-    ObjectPosSelector selector(GetPositionX(), GetPositionY(), distance2d, searcher_bounding_radius, searcher);
-
-    // adding used positions around object
-    {
-        MaNGOS::NearUsedPosDo u_do(*this, searcher, absAngle, selector);
-        MaNGOS::WorldObjectWorker<MaNGOS::NearUsedPosDo> worker(this, u_do);
-
-        Cell::VisitAllObjects(this, worker, dist);
-    }
-
-    // maybe can just place in primary position
-    if (selector.CheckOriginalAngle())
-    {
-        if (searcher)
-        {
-            searcher->UpdateAllowedPositionZ(x, y, z, GetMap());       // update to LOS height if available
-        }
-        else
-        {
-            UpdateGroundPositionZ(x, y, z);
-        }
-
-        if (fabs(init_z - z) < dist && IsWithinLOS(x, y, z))
-        {
-            return;
-        }
-
-        first_los_conflict = true;                          // first point have LOS problems
-    }
-
-    // set first used pos in lists
-    selector.InitializeAngle();
-
-    float angle;                                            // candidate of angle for free pos
-
-    // select in positions after current nodes (selection one by one)
-    while (selector.NextAngle(angle))                       // angle for free pos
-    {
-        GetNearPoint2D(x, y, distance2d, absAngle + angle);
-        z = GetPositionZ();
-
-        if (searcher)
-        {
-            searcher->UpdateAllowedPositionZ(x, y, z, GetMap());       // update to LOS height if available
-        }
-        else
-        {
-            UpdateGroundPositionZ(x, y, z);
-        }
-
-        if (fabs(init_z - z) < dist && IsWithinLOS(x, y, z))
-        {
-            return;
-        }
-    }
-
-    // BAD NEWS: not free pos (or used or have LOS problems)
-    // Attempt find _used_ pos without LOS problem
-    if (!first_los_conflict)
-    {
-        x = first_x;
-        y = first_y;
-
-        if (searcher)
-        {
-            searcher->UpdateAllowedPositionZ(x, y, z, GetMap());       // update to LOS height if available
-        }
-        else
-        {
-            UpdateGroundPositionZ(x, y, z);
-        }
-        return;
-    }
-
-    // set first used pos in lists
-    selector.InitializeAngle();
-
-    // select in positions after current nodes (selection one by one)
-    while (selector.NextUsedAngle(angle))                   // angle for used pos but maybe without LOS problem
-    {
-        GetNearPoint2D(x, y, distance2d, absAngle + angle);
-        z = GetPositionZ();
-
-        if (searcher)
-        {
-            searcher->UpdateAllowedPositionZ(x, y, z, GetMap());       // update to LOS height if available
-        }
-        else
-        {
-            UpdateGroundPositionZ(x, y, z);
-        }
-
-        if (fabs(init_z - z) < dist && IsWithinLOS(x, y, z))
-        {
-            return;
-        }
-    }
-
-    // BAD BAD NEWS: all found pos (free and used) have LOS problem :(
-    x = first_x;
-    y = first_y;
-
-    if (searcher)
-    {
-        searcher->UpdateAllowedPositionZ(x, y, z, GetMap());           // update to LOS height if available
-    }
-    else
-    {
-        UpdateGroundPositionZ(x, y, z);
-    }
-}
-
-void WorldObject::SetPhaseMask(uint32 newPhaseMask, bool update)
-{
-    m_phaseMask = newPhaseMask;
-
-    if (update && IsInWorld())
-    {
-        UpdateVisibilityAndView();
-    }
-}
-
-/**
- * @brief Plays a positional sound for one player or nearby players.
- *
- * @param sound_id The sound entry id.
- * @param target Optional single-player target.
- */
-void WorldObject::PlayDistanceSound(uint32 sound_id, Player const* target /*= NULL*/) const
-{
-    WorldPacket data(SMSG_PLAY_OBJECT_SOUND, 4 + 8);
-    data << uint32(sound_id);
-    data << GetObjectGuid();
-    data << GetObjectGuid();
-    if (target)
-    {
-        target->SendDirectMessage(&data);
-    }
-    else
-    {
-        SendMessageToSet(&data, true);
-    }
-}
-
-/**
- * @brief Plays a direct sound for one player or nearby players.
- *
- * @param sound_id The sound entry id.
- * @param target Optional single-player target.
- */
-void WorldObject::PlayDirectSound(uint32 sound_id, Player const* target /*= NULL*/) const
-{
-    WorldPacket data(SMSG_PLAY_SOUND, 4);
-    data << uint32(sound_id);
-    data << ObjectGuid();
-    if (target)
-    {
-        target->SendDirectMessage(&data);
-    }
-    else
-    {
-        SendMessageToSet(&data, true);
-    }
-}
-
-/**
- * @brief Plays music for one player or nearby players.
- *
- * @param sound_id The music entry id.
- * @param target Optional single-player target.
- */
-void WorldObject::PlayMusic(uint32 sound_id, Player const* target /*= NULL*/) const
-{
-    WorldPacket data(SMSG_PLAY_MUSIC, 4);
-    data << uint32(sound_id);
-    if (target)
-    {
-        target->SendDirectMessage(&data);
-    }
-    else
-    {
-        SendMessageToSet(&data, true);
-    }
-}
-
-/**
- * @brief Refreshes both visibility and viewpoint-dependent visibility state.
- */
-void WorldObject::UpdateVisibilityAndView()
-{
-    GetViewPoint().Call_UpdateVisibilityForOwner();
-    UpdateObjectVisibility();
-    GetViewPoint().Event_ViewPointVisibilityChanged();
-}
-
-/**
- * @brief Recomputes this object's visibility for nearby clients.
- */
-void WorldObject::UpdateObjectVisibility()
-{
-    CellPair p = MaNGOS::ComputeCellPair(GetPositionX(), GetPositionY());
-    Cell cell(p);
-
-    GetMap()->UpdateObjectVisibility(this, cell, p);
-}
-
-/**
- * @brief Adds the world object to the map's update queue.
- */
-void WorldObject::AddToClientUpdateList()
-{
-    GetMap()->AddUpdateObject(this);
-}
-
-/**
- * @brief Remove from client update list
- *
- * Removes this object from the map's update list.
- */
-void WorldObject::RemoveFromClientUpdateList()
-{
-    GetMap()->RemoveUpdateObject(this);
-}
-
-/**
- * @brief World object change accumulator
- *
- * Accumulates update data for a world object and nearby players.
- */
-struct WorldObjectChangeAccumulator
-{
-    UpdateDataMapType& i_updateDatas; ///< Update data map
-    WorldObject& i_object; ///< World object
-
-    /**
-     * @brief Constructor
-     * @param obj World object
-     * @param d Update data map
-     */
-    WorldObjectChangeAccumulator(WorldObject& obj, UpdateDataMapType& d) : i_updateDatas(d), i_object(obj)
-    {
-        // send self fields changes in another way, otherwise
-        // with new camera system when player's camera too far from player, camera wouldn't receive packets and changes from player
-        if (i_object.isType(TYPEMASK_PLAYER))
-        {
-            i_object.BuildUpdateDataForPlayer((Player*)&i_object, i_updateDatas);
-        }
-    }
-
-    /**
-     * @brief Visit cameras
-     * @param m Camera map
-     *
-     * Builds update data for all camera owners that can see this object.
-     */
-    void Visit(CameraMapType& m)
-    {
-        for (CameraMapType::iterator iter = m.begin(); iter != m.end(); ++iter)
-        {
-            Player* owner = iter->getSource()->GetOwner();
-            if (owner != &i_object && owner->HaveAtClient(&i_object))
-            {
-                i_object.BuildUpdateDataForPlayer(owner, i_updateDatas);
-            }
-        }
-    }
-
-    /**
-     * @brief Visit other grid references (no-op)
-     */
-    template<class SKIP> void Visit(GridRefManager<SKIP>&) {}
-};
-
-/**
- * @brief Build update data
- * @param update_players Map of players to their update data
- *
- * Builds update data for all players who can see this object.
- */
-void WorldObject::BuildUpdateData(UpdateDataMapType& update_players)
-{
-    WorldObjectChangeAccumulator notifier(*this, update_players);
-    Cell::VisitWorldObjects(this, notifier, GetMap()->GetBroadcastRadius());
-
-    ClearUpdateMask(false);
-}
-
-bool WorldObject::IsControlledByPlayer() const
-{
-    switch (GetTypeId())
-    {
-        case TYPEID_GAMEOBJECT:
-            return ((GameObject*)this)->GetOwnerGuid().IsPlayer();
-        case TYPEID_UNIT:
-        case TYPEID_PLAYER:
-            return ((Unit*)this)->IsCharmerOrOwnerPlayerOrPlayerItself();
-        case TYPEID_DYNAMICOBJECT:
-            return ((DynamicObject*)this)->GetCasterGuid().IsPlayer();
-        case TYPEID_CORPSE:
-            return true;
-        default:
-            return false;
-    }
-}
-
-/**
- * @brief Print coordinates error
- * @param x X coordinate
- * @param y Y coordinate
- * @param z Z coordinate
- * @param descr Description of the operation
- * @return Always false
- *
- * Logs an error when invalid coordinates are encountered.
- */
-bool WorldObject::PrintCoordinatesError(float x, float y, float z, char const* descr) const
-{
-    sLog.outError("%s with invalid %s coordinates: mapid = %uu, x = %f, y = %f, z = %f", GetGuidStr().c_str(), descr, GetMapId(), x, y, z);
-    return false;                                           // always false for continue assert fail
-}
-
-/**
- * @brief Set active object state
- * @param active If true, set as active object
- *
- * Sets whether this object is an active object (updated even when no players nearby).
- */
-void WorldObject::SetActiveObjectState(bool active)
-{
-    if (m_isActiveObject == active || (isType(TYPEMASK_PLAYER) && !active))  // player shouldn't became inactive, never
-    {
-        return;
-    }
-
-    if (IsInWorld() && !isType(TYPEMASK_PLAYER))
-        // player's update implemented in a different from other active worldobject's way
-        // considered using a generic way in future
-    {
-        if (IsActiveObject() && !active)
-        {
-            GetMap()->RemoveFromActive(this);
-        }
-        else if (IsActiveObject() && active)
-        {
-            GetMap()->AddToActive(this);
-        }
-    }
-    m_isActiveObject = active;
-}
-
-#ifdef ENABLE_ELUNA
-/**
- * @brief Get Eluna instance
- * @return Eluna instance pointer or nullptr
- *
- * Returns the Eluna scripting engine instance for this object's map.
- */
-Eluna* WorldObject::GetEluna() const
-{
-    if (IsInWorld())
-    {
-        return GetMap()->GetEluna();
-    }
-
-    return nullptr;
-}
-#endif /* ENABLE_ELUNA */
