+++ src/game/Object/ObjectUpdate.cpp
@@ -1,0 +1,937 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+#include "Object.h"
+#include "SharedDefines.h"
+#include "WorldPacket.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "World.h"
+#include "Creature.h"
+#include "Player.h"
+#include "Vehicle.h"
+#include "ObjectMgr.h"
+#include "ObjectGuid.h"
+#include "UpdateData.h"
+#include "UpdateMask.h"
+#include "Util.h"
+#include "MapManager.h"
+#include "Log.h"
+#include "Transports.h"
+#include "TargetedMovementGenerator.h"
+#include "WaypointMovementGenerator.h"
+#include "VMapFactory.h"
+#include "CellImpl.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "ObjectPosSelector.h"
+#include "TemporarySummon.h"
+#include "movement/packet_builder.h"
+#include "CreatureLinkingMgr.h"
+#include "Chat.h"
+#include "GameTime.h"
+
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#include "ElunaConfig.h"
+#include "ElunaEventMgr.h"
+#endif /* ENABLE_ELUNA */
+
+/**
+ * @file ObjectUpdate.cpp
+ * @brief Cohesion split of Object.cpp -- Object update-data serialization: create/values update-block building, movement-block packing, update masks and value (de)serialization. Same Object class; no behaviour change. CMake file(GLOB Object/*.cpp) picks this file up automatically; Object.h is unchanged.
+ */
+
+/**
+ * @brief Force immediate update transmission to all viewers
+ *
+ * Sends all pending update changes immediately rather than waiting
+ * for the next update tick. This is used for urgent updates that
+ * must be visible immediately (e.g., combat state changes).
+ *
+ * The method builds update data for all nearby players and sends
+ * it immediately, then removes the object from the pending update list.
+ */
+void Object::SendForcedObjectUpdate()
+{
+    if (!m_inWorld || !m_objectUpdated)
+    {
+        return;
+    }
+
+    UpdateDataMapType update_players;
+
+    BuildUpdateData(update_players);
+    RemoveFromClientUpdateList();
+
+    WorldPacket packet;                                     // here we allocate a std::vector with a size of 0x10000
+    for (UpdateDataMapType::iterator iter = update_players.begin(); iter != update_players.end(); ++iter)
+    {
+        iter->second.BuildPacket(&packet);
+        iter->first->GetSession()->SendPacket(&packet);
+        packet.clear();                                     // clean the string
+    }
+}
+
+/**
+ * @brief Build create update block for player
+ * @param data Update data buffer
+ * @param target Target player
+ *
+ * Builds the update packet data needed to create this object
+ * for the specified player. Includes movement data and
+ * all update field values.
+ */
+void Object::BuildCreateUpdateBlockForPlayer(UpdateData* data, Player* target) const
+{
+    if (!target)
+    {
+        return;
+    }
+
+    uint8  updatetype   = UPDATETYPE_CREATE_OBJECT;
+    uint16 updateFlags  = m_updateFlag;
+
+    /** lower flag1 **/
+    if (target == this)                                     // building packet for yourself
+    {
+        updateFlags |= UPDATEFLAG_SELF;
+    }
+
+    if (m_itsNewObject)
+    {
+        switch (GetObjectGuid().GetHigh())
+        {
+            case HighGuid::HIGHGUID_DYNAMICOBJECT:
+            case HighGuid::HIGHGUID_CORPSE:
+            case HighGuid::HIGHGUID_PLAYER:
+            case HighGuid::HIGHGUID_UNIT:
+            case HighGuid::HIGHGUID_VEHICLE:
+            case HighGuid::HIGHGUID_GAMEOBJECT:
+                updatetype = UPDATETYPE_CREATE_OBJECT2;
+                break;
+
+            default:
+                break;
+        }
+    }
+
+    if (isType(TYPEMASK_UNIT))
+    {
+        if (((Unit*)this)->getVictim())
+        {
+            updateFlags |= UPDATEFLAG_HAS_ATTACKING_TARGET;
+        }
+    }
+
+    // DEBUG_LOG("BuildCreateUpdate: update-type: %u, object-type: %u got updateFlags: %X", updatetype, m_objectTypeId, updateFlags);
+
+    ByteBuffer& buf = data->GetBuffer();
+    buf << uint8(updatetype);
+    buf << GetPackGUID();
+    buf << uint8(m_objectTypeId);
+
+    BuildMovementUpdate(&buf, updateFlags);
+
+    UpdateMask updateMask;
+    updateMask.SetCount(m_valuesCount);
+    _SetCreateBits(&updateMask, target);
+    BuildValuesUpdate(updatetype, &buf, &updateMask, target);
+    data->AddUpdateBlock();
+}
+
+/**
+ * @brief Send create update to player
+ * @param player Target player
+ *
+ * Sends the create update packet to the specified player,
+ * causing the object to appear in their game world.
+ */
+void Object::SendCreateUpdateToPlayer(Player* player)
+{
+    // send create update to player
+    UpdateData upd(player->GetMapId());
+    WorldPacket packet;
+
+    BuildCreateUpdateBlockForPlayer(&upd, player);
+    upd.BuildPacket(&packet);
+    player->GetSession()->SendPacket(&packet);
+}
+
+/**
+ * @brief Build values update block for player
+ * @param data Update data buffer
+ * @param target Target player
+ *
+ * Builds the update packet data for changed field values
+ * to send to the specified player.
+ */
+void Object::BuildValuesUpdateBlockForPlayer(UpdateData* data, Player* target) const
+{
+    ByteBuffer& buf = data->GetBuffer();
+
+    buf << uint8(UPDATETYPE_VALUES);
+    buf << GetPackGUID();
+
+    UpdateMask updateMask;
+    updateMask.SetCount(m_valuesCount);
+
+    _SetUpdateBits(&updateMask, target);
+    BuildValuesUpdate(UPDATETYPE_VALUES, &buf, &updateMask, target);
+
+    data->AddUpdateBlock();
+}
+
+/**
+ * @brief Build out of range update block
+ * @param data Update data buffer
+ *
+ * Adds this object's GUID to the out-of-range list,
+ * indicating it should be removed from the client's view.
+ */
+void Object::BuildOutOfRangeUpdateBlock(UpdateData* data) const
+{
+    data->AddOutOfRangeGUID(GetObjectGuid());
+}
+
+/**
+ * @brief Destroy object for player
+ * @param target Target player
+ *
+ * Sends a destroy packet to the specified player,
+ * removing this object from their game world.
+ */
+void Object::DestroyForPlayer(Player* target, bool anim) const
+{
+    MANGOS_ASSERT(target);
+
+    WorldPacket data(SMSG_DESTROY_OBJECT, 9);
+    data << GetObjectGuid();
+    data << uint8(anim ? 1 : 0);                            // WotLK (bool), may be despawn animation
+    target->GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Build movement update block
+ * @param data Byte buffer to write to
+ * @param updateFlags Update flags
+ *
+ * Builds the movement data portion of the update packet.
+ * Includes position, orientation, movement flags, and speeds
+ * for living objects, or just position for static objects.
+ */
+void Object::BuildMovementUpdate(ByteBuffer* data, uint16 updateFlags) const
+{
+    ObjectGuid Guid = GetObjectGuid();
+
+    data->WriteBit(false);
+    data->WriteBit(false);
+    data->WriteBit(updateFlags & UPDATEFLAG_ROTATION);
+    data->WriteBit(updateFlags & UPDATEFLAG_ANIM_KITS);               // AnimKits
+    data->WriteBit(updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET);
+    data->WriteBit(updateFlags & UPDATEFLAG_SELF);
+    data->WriteBit(updateFlags & UPDATEFLAG_VEHICLE);
+    data->WriteBit(updateFlags & UPDATEFLAG_LIVING);
+    data->WriteBits(0, 24);                                     // Byte Counter
+    data->WriteBit(false);
+    data->WriteBit(updateFlags & UPDATEFLAG_POSITION);                // flags & UPDATEFLAG_HAS_POSITION Game Object Position
+    data->WriteBit(updateFlags & UPDATEFLAG_HAS_POSITION);            // Stationary Position
+    data->WriteBit(updateFlags & UPDATEFLAG_TRANSPORT_ARR);
+    data->WriteBit(false);
+    data->WriteBit(updateFlags & UPDATEFLAG_TRANSPORT);
+
+    bool hasTransport = false,
+        isSplineEnabled = false,
+        hasPitch = false,
+        hasFallData = false,
+        hasFallDirection = false,
+        hasElevation = false,
+        hasOrientation = !isType(TYPEMASK_ITEM),
+        hasTimeStamp = true,
+        hasTransportTime2 = false,
+        hasTransportTime3 = false;
+
+    if (isType(TYPEMASK_UNIT))
+    {
+        Unit const* unit = (Unit const*)this;
+        hasTransport = !unit->m_movementInfo.GetTransportGuid().IsEmpty();
+        isSplineEnabled = unit->IsSplineEnabled();
+
+        if (GetTypeId() == TYPEID_PLAYER)
+        {
+            // use flags received from client as they are more correct
+            hasPitch = unit->m_movementInfo.GetStatusInfo().hasPitch;
+            hasFallData = unit->m_movementInfo.GetStatusInfo().hasFallData;
+            hasFallDirection = unit->m_movementInfo.GetStatusInfo().hasFallDirection;
+            hasElevation = unit->m_movementInfo.GetStatusInfo().hasSplineElevation;
+            hasTransportTime2 = unit->m_movementInfo.GetStatusInfo().hasTransportTime2;
+            hasTransportTime3 = unit->m_movementInfo.GetStatusInfo().hasTransportTime3;
+        }
+        else
+        {
+            hasPitch = unit->m_movementInfo.HasMovementFlag(MovementFlags(MOVEFLAG_SWIMMING | MOVEFLAG_FLYING)) ||
+                            unit->m_movementInfo.HasMovementFlag2(MOVEFLAG2_ALLOW_PITCHING);
+            hasFallData = unit->m_movementInfo.HasMovementFlag2(MOVEFLAG2_INTERP_TURNING);
+            hasFallDirection = unit->m_movementInfo.HasMovementFlag(MOVEFLAG_FALLING);
+            hasElevation = unit->m_movementInfo.HasMovementFlag(MOVEFLAG_SPLINE_ELEVATION);
+        }
+    }
+
+    if (updateFlags & UPDATEFLAG_LIVING)
+    {
+        Unit const* unit = (Unit const*)this;
+
+        data->WriteBit(!unit->m_movementInfo.GetMovementFlags());
+        data->WriteBit(!hasOrientation);
+
+        data->WriteGuidMask<7, 3, 2>(Guid);
+
+        if (unit->m_movementInfo.GetMovementFlags())
+        {
+            data->WriteBits(unit->m_movementInfo.GetMovementFlags(), 30);
+        }
+
+        data->WriteBit(false);
+        data->WriteBit(!hasPitch);
+        data->WriteBit(isSplineEnabled);
+        data->WriteBit(hasFallData);
+        data->WriteBit(!hasElevation);
+        data->WriteGuidMask<5>(Guid);
+        data->WriteBit(hasTransport);
+        data->WriteBit(!hasTimeStamp);
+
+        if (hasTransport)
+        {
+            ObjectGuid tGuid = unit->m_movementInfo.GetTransportGuid();
+
+            data->WriteGuidMask<1>(tGuid);
+            data->WriteBit(hasTransportTime2);
+            data->WriteGuidMask<4, 0, 6>(tGuid);
+            data->WriteBit(hasTransportTime3);
+            data->WriteGuidMask<7, 5, 3, 2>(tGuid);
+        }
+
+        data->WriteGuidMask<4>(Guid);
+
+        if (isSplineEnabled)
+        {
+            Movement::PacketBuilder::WriteCreateBits(*unit->movespline, *data);
+        }
+
+        data->WriteGuidMask<6>(Guid);
+
+        if (hasFallData)
+        {
+            data->WriteBit(hasFallDirection);
+        }
+
+        data->WriteGuidMask<0, 1>(Guid);
+        data->WriteBit(false);    // Unknown 4.3.3
+        data->WriteBit(!unit->m_movementInfo.GetMovementFlags2());
+
+        if (unit->m_movementInfo.GetMovementFlags2())
+        {
+            data->WriteBits(unit->m_movementInfo.GetMovementFlags2(), 12);
+        }
+    }
+
+    // used only with GO's, placeholder
+    if (updateFlags & UPDATEFLAG_POSITION)
+    {
+        ObjectGuid transGuid;
+        data->WriteGuidMask<5>(transGuid);
+        data->WriteBit(hasTransportTime3);
+        data->WriteGuidMask<0, 3, 6, 1, 4, 2>(transGuid);
+        data->WriteBit(hasTransportTime2);
+        data->WriteGuidMask<7>(transGuid);
+    }
+
+    if (updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET)
+    {
+        ObjectGuid guid;
+        if (Unit* victim = ((Unit*)this)->getVictim())
+        {
+            guid = victim->GetObjectGuid();
+        }
+
+        data->WriteGuidMask<2, 7, 0, 4, 5, 6, 1, 3>(guid);
+    }
+
+    if (updateFlags & UPDATEFLAG_ANIM_KITS)
+    {
+        data->WriteBit(true);   // hasAnimKit0 == false
+        data->WriteBit(true);   // hasAnimKit1 == false
+        data->WriteBit(true);   // hasAnimKit2 == false
+    }
+
+    data->FlushBits();
+
+    if (updateFlags & UPDATEFLAG_LIVING)
+    {
+        Unit const* unit = (Unit const*)this;
+
+        data->WriteGuidBytes<4>(Guid);
+
+        *data << float(unit->GetSpeed(MOVE_RUN_BACK));
+
+        if (hasFallData)
+        {
+            if (hasFallDirection)
+            {
+                // 15595 client reads horizontal speed, then jump direction sin/cos
+                *data << float(unit->m_movementInfo.GetJumpInfo().xyspeed);
+                *data << float(unit->m_movementInfo.GetJumpInfo().sinAngle);
+                *data << float(unit->m_movementInfo.GetJumpInfo().cosAngle);
+            }
+
+            *data << uint32(unit->m_movementInfo.GetFallTime());
+            *data << float(unit->m_movementInfo.GetJumpInfo().velocity);
+        }
+
+        *data << float(unit->GetSpeed(MOVE_SWIM_BACK));
+
+        if (hasElevation)
+        {
+            *data << float(unit->m_movementInfo.GetSplineElevation());
+        }
+
+        if (isSplineEnabled)
+        {
+            Movement::PacketBuilder::WriteCreateBytes(*unit->movespline, *data);
+        }
+
+        *data << float(unit->GetPositionZ());
+        data->WriteGuidBytes<5>(Guid);
+
+        if (hasTransport)
+        {
+            ObjectGuid tGuid = unit->m_movementInfo.GetTransportGuid();
+
+            data->WriteGuidBytes<5, 7>(tGuid);
+            *data << uint32(unit->m_movementInfo.GetTransportTime());
+            *data << float(NormalizeOrientation(unit->m_movementInfo.GetTransportPos()->o));
+
+            if (hasTransportTime2)
+            {
+                *data << uint32(unit->m_movementInfo.GetTransportTime2());
+            }
+
+            *data << float(unit->m_movementInfo.GetTransportPos()->y);
+            *data << float(unit->m_movementInfo.GetTransportPos()->x);
+            data->WriteGuidBytes<3>(tGuid);
+            *data << float(unit->m_movementInfo.GetTransportPos()->z);
+            data->WriteGuidBytes<0>(tGuid);
+
+            if (hasTransportTime3)
+            {
+                *data << uint32(unit->m_movementInfo.GetFallTime());
+            }
+
+            *data << int8(unit->m_movementInfo.GetTransportSeat());
+            data->WriteGuidBytes<1, 6, 2, 4>(tGuid);
+        }
+
+        *data << float(unit->GetPositionX());
+        *data << float(unit->GetSpeed(MOVE_PITCH_RATE));
+        data->WriteGuidBytes<3, 0>(Guid);
+        *data << float(unit->GetSpeed(MOVE_SWIM));
+        *data << float(unit->GetPositionY());
+        data->WriteGuidBytes<7, 1, 2>(Guid);
+        *data << float(unit->GetSpeed(MOVE_WALK));
+
+        *data << uint32(GameTime::GetGameTimeMS());
+
+        *data << float(unit->GetSpeed(MOVE_FLIGHT_BACK));
+        data->WriteGuidBytes<6>(Guid);
+        *data << float(unit->GetSpeed(MOVE_TURN_RATE));
+
+        if (hasOrientation)
+        {
+            *data << float(NormalizeOrientation(unit->GetOrientation()));
+        }
+
+        *data << float(unit->GetSpeed(MOVE_RUN));
+
+        if (hasPitch)
+        {
+            *data << float(unit->m_movementInfo.GetPitch());
+        }
+
+        *data << float(unit->GetSpeed(MOVE_FLIGHT));
+    }
+
+    if (updateFlags & UPDATEFLAG_VEHICLE)
+    {
+        *data << float(NormalizeOrientation(((WorldObject*)this)->GetOrientation()));
+        *data << uint32(((Unit*)this)->GetVehicleInfo()->GetVehicleEntry()->m_ID); // vehicle id
+    }
+
+    // used only with GO's, placeholder
+    if (updateFlags & UPDATEFLAG_POSITION)
+    {
+        ObjectGuid transGuid;
+
+        data->WriteGuidBytes<0, 5>(transGuid);
+        if (hasTransportTime3)
+        {
+            *data << uint32(0);
+        }
+
+        data->WriteGuidBytes<3>(transGuid);
+        *data << float(0.0f);   // x offset
+        data->WriteGuidBytes<4, 6, 1>(transGuid);
+        *data << uint32(0);     // transport time
+        *data << float(0.0f);   // y offset
+        data->WriteGuidBytes<2, 7>(transGuid);
+        *data << float(0.0f);   // z offset
+        *data << int8(-1);      // transport seat
+        *data << float(0.0f);   // o offset
+
+        if (hasTransportTime2)
+        {
+            *data << uint32(0);
+        }
+    }
+
+    if (updateFlags & UPDATEFLAG_ROTATION)
+    {
+        *data << int64(((GameObject*)this)->GetPackedWorldRotation());
+    }
+
+    if (updateFlags & UPDATEFLAG_TRANSPORT_ARR)
+    {
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << uint8(0);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+        *data << float(0.0f);
+    }
+
+    if (updateFlags & UPDATEFLAG_HAS_POSITION)
+    {
+        *data << float(NormalizeOrientation(((WorldObject*)this)->GetOrientation()));
+        *data << float(((WorldObject*)this)->GetPositionX());
+        *data << float(((WorldObject*)this)->GetPositionY());
+        *data << float(((WorldObject*)this)->GetPositionZ());
+    }
+
+    if (updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET)
+    {
+        ObjectGuid guid;
+        if (Unit* victim = ((Unit*)this)->getVictim())
+        {
+            guid = victim->GetObjectGuid();
+        }
+
+        data->WriteGuidBytes<4, 0, 3, 5, 7, 6, 2, 1>(guid);
+    }
+
+    if (updateFlags & UPDATEFLAG_TRANSPORT)
+    {
+        *data << uint32(GameTime::GetGameTimeMS());           // ms time
+    }
+}
+
+/**
+ * @brief Build values update data
+ * @param updatetype Update type (create or values)
+ * @param data Byte buffer to write to
+ * @param updateMask Update mask indicating which fields changed
+ * @param target Target player
+ *
+ * Builds the actual field value data for the update packet.
+ * Handles special cases for gameobjects and units.
+ */
+void Object::BuildValuesUpdate(uint8 updatetype, ByteBuffer* data, UpdateMask* updateMask, Player* target) const
+{
+    if (!target)
+    {
+        return;
+    }
+
+    uint32 valuesCount = m_valuesCount;
+    if (GetTypeId() == TYPEID_PLAYER && target != this)
+    {
+        valuesCount = PLAYER_END_NOT_SELF;
+    }
+
+    bool IsActivateToQuest = false;
+    bool IsPerCasterAuraState = false;
+
+    if (updatetype == UPDATETYPE_CREATE_OBJECT || updatetype == UPDATETYPE_CREATE_OBJECT2)
+    {
+        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
+        {
+            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
+            {
+                IsActivateToQuest = true;
+            }
+
+            updateMask->SetBit(GAMEOBJECT_DYNAMIC);
+        }
+        else if (isType(TYPEMASK_UNIT))
+        {
+            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
+            {
+                IsPerCasterAuraState = true;
+                updateMask->SetBit(UNIT_FIELD_AURASTATE);
+            }
+        }
+    }
+    else                                                    // case UPDATETYPE_VALUES
+    {
+        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
+        {
+            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
+            {
+                IsActivateToQuest = true;
+            }
+
+            updateMask->SetBit(GAMEOBJECT_DYNAMIC);
+            updateMask->SetBit(GAMEOBJECT_BYTES_1);         // why do we need this here?
+        }
+        else if (isType(TYPEMASK_UNIT))
+        {
+            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
+            {
+                IsPerCasterAuraState = true;
+                updateMask->SetBit(UNIT_FIELD_AURASTATE);
+            }
+        }
+    }
+
+    MANGOS_ASSERT(updateMask && updateMask->GetCount() == m_valuesCount);
+
+    *data << (uint8)updateMask->GetBlockCount();
+    data->append(updateMask->GetMask(), updateMask->GetLength());
+
+    // 2 specialized loops for speed optimization in non-unit case
+    if (isType(TYPEMASK_UNIT))                              // unit (creature/player) case
+    {
+        for (uint16 index = 0; index < valuesCount; ++index)
+        {
+            if (updateMask->GetBit(index))
+            {
+                if (index == UNIT_NPC_FLAGS)
+                {
+                    uint32 appendValue = m_uint32Values[index];
+
+                    if (GetTypeId() == TYPEID_UNIT)
+                    {
+                        if (!target->canSeeSpellClickOn((Creature*)this))
+                        {
+                            appendValue &= ~UNIT_NPC_FLAG_SPELLCLICK;
+                        }
+
+                        if (appendValue & UNIT_NPC_FLAG_TRAINER)
+                        {
+                            if (!((Creature*)this)->IsTrainerOf(target, false))
+                            {
+                                appendValue &= ~(UNIT_NPC_FLAG_TRAINER | UNIT_NPC_FLAG_TRAINER_CLASS | UNIT_NPC_FLAG_TRAINER_PROFESSION);
+                            }
+                        }
+
+                        if (appendValue & UNIT_NPC_FLAG_STABLEMASTER)
+                        {
+                            if (target->getClass() != CLASS_HUNTER)
+                            {
+                                appendValue &= ~UNIT_NPC_FLAG_STABLEMASTER;
+                            }
+                        }
+                    }
+
+                    *data << uint32(appendValue);
+                }
+                else if (index == UNIT_FIELD_AURASTATE)
+                {
+                    if (IsPerCasterAuraState)
+                    {
+                        // IsPerCasterAuraState set if related pet caster aura state set already
+                        if (((Unit*)this)->HasAuraStateForCaster(AURA_STATE_CONFLAGRATE, target->GetObjectGuid()))
+                        {
+                            *data << m_uint32Values[index];
+                        }
+                        else
+                        {
+                            *data << (m_uint32Values[index] & ~(1 << (AURA_STATE_CONFLAGRATE - 1)));
+                        }
+                    }
+                    else
+                    {
+                        *data << m_uint32Values[index];
+                    }
+                }
+                // FIXME: Some values at server stored in float format but must be sent to client in uint32 format
+                else if (index >= UNIT_FIELD_BASEATTACKTIME && index <= UNIT_FIELD_RANGEDATTACKTIME)
+                {
+                    // convert from float to uint32 and send
+                    *data << uint32(m_floatValues[index] < 0 ? 0 : m_floatValues[index]);
+                }
+
+                // there are some float values which may be negative or can't get negative due to other checks
+                else if ((index >= UNIT_FIELD_NEGSTAT0 && index <= UNIT_FIELD_NEGSTAT4) ||
+                         (index >= UNIT_FIELD_RESISTANCEBUFFMODSPOSITIVE  && index <= (UNIT_FIELD_RESISTANCEBUFFMODSPOSITIVE + 6)) ||
+                         (index >= UNIT_FIELD_RESISTANCEBUFFMODSNEGATIVE  && index <= (UNIT_FIELD_RESISTANCEBUFFMODSNEGATIVE + 6)) ||
+                         (index >= UNIT_FIELD_POSSTAT0 && index <= UNIT_FIELD_POSSTAT4))
+                {
+                    *data << uint32(m_floatValues[index]);
+                }
+
+                // Gamemasters should be always able to select units - remove not selectable flag
+                else if (index == UNIT_FIELD_FLAGS && target->isGameMaster())
+                {
+                    *data << (m_uint32Values[index] & ~UNIT_FLAG_NOT_SELECTABLE);
+                }
+                /* Hide loot animation for players that aren't permitted to loot the corpse */
+                else if (index == UNIT_DYNAMIC_FLAGS && GetTypeId() == TYPEID_UNIT)
+                {
+                    uint32 send_value = m_uint32Values[index];
+
+                    /* Initiate pointer to creature so we can check loot */
+                    if (Creature* my_creature = (Creature*)this)
+                        /* If the creature is NOT fully looted */
+                        if (!my_creature->loot.isLooted())
+                            /* If the lootable flag is NOT set */
+                            if (!(send_value & UNIT_DYNFLAG_LOOTABLE))
+                            {
+                                /* Update it on the creature */
+                                my_creature->SetFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_LOOTABLE);
+                                /* Update it in the packet */
+                                send_value = send_value | UNIT_DYNFLAG_LOOTABLE;
+                            }
+
+                    /* If we're not allowed to loot the target, destroy the lootable flag */
+                    if (!target->isAllowedToLoot((Creature*)this))
+                        if (send_value & UNIT_DYNFLAG_LOOTABLE)
+                        {
+                            send_value = send_value & ~UNIT_DYNFLAG_LOOTABLE;
+                        }
+
+                    /* If we are allowed to loot it and mob is tapped by us, destroy the tapped flag */
+                    bool is_tapped = target->IsTappedByMeOrMyGroup((Creature*)this);
+
+                    /* If the creature has tapped flag but is tapped by us, remove the flag */
+                    if (send_value & UNIT_DYNFLAG_TAPPED && is_tapped)
+                    {
+                        send_value = send_value & ~UNIT_DYNFLAG_TAPPED;
+                    }
+
+                    *data << send_value;
+                }
+                else
+                {
+                    // send in current format (float as float, uint32 as uint32)
+                    *data << m_uint32Values[index];
+                }
+            }
+        }
+    }
+    else if (isType(TYPEMASK_GAMEOBJECT))                   // gameobject case
+    {
+        for (uint16 index = 0; index < valuesCount; ++index)
+        {
+            if (updateMask->GetBit(index))
+            {
+                // send in current format (float as float, uint32 as uint32)
+                if (index == GAMEOBJECT_DYNAMIC)
+                {
+                    // GAMEOBJECT_TYPE_DUNGEON_DIFFICULTY can have lo flag = 2
+                    //      most likely related to "can enter map" and then should be 0 if can not enter
+
+                    if (IsActivateToQuest)
+                    {
+                        switch (((GameObject*)this)->GetGoType())
+                        {
+                            case GAMEOBJECT_TYPE_QUESTGIVER:
+                                // GO also seen with GO_DYNFLAG_LO_SPARKLE explicit, relation/reason unclear (192861)
+                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE);
+                                *data << uint16(-1);
+                                break;
+                            case GAMEOBJECT_TYPE_CHEST:
+                            case GAMEOBJECT_TYPE_GENERIC:
+                            case GAMEOBJECT_TYPE_SPELL_FOCUS:
+                            case GAMEOBJECT_TYPE_GOOBER:
+                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE | GO_DYNFLAG_LO_SPARKLE);
+                                *data << uint16(-1);
+                                break;
+                            default:
+                                // unknown, not happen.
+                                *data << uint16(0);
+                                *data << uint16(-1);
+                                break;
+                        }
+                    }
+                    else
+                    {
+                        // disable quest object
+                        *data << uint16(0);
+                        *data << uint16(-1);
+                    }
+                }
+                else if (index == GAMEOBJECT_BYTES_1)
+                {
+                    if (((GameObject*)this)->GetGOInfo()->type == GAMEOBJECT_TYPE_TRANSPORT)
+                    {
+                        *data << uint32(m_uint32Values[index] | GO_STATE_TRANSPORT_SPEC);
+                    }
+                    else
+                    {
+                        *data << uint32(m_uint32Values[index]);
+                    }
+                }
+                else
+                {
+                    *data << m_uint32Values[index];          // other cases
+                }
+            }
+        }
+    }
+    else                                                    // other objects case (no special index checks)
+    {
+        for (uint16 index = 0; index < valuesCount; ++index)
+        {
+            if (updateMask->GetBit(index))
+            {
+                // send in current format (float as float, uint32 as uint32)
+                *data << m_uint32Values[index];
+            }
+        }
+    }
+}
+
+/**
+ * @brief Clear update mask
+ * @param remove If true, remove from client update list
+ *
+ * Clears all changed value flags and optionally removes
+ * the object from the pending update list.
+ */
+void Object::ClearUpdateMask(bool remove)
+{
+    if (m_uint32Values)
+    {
+        for (uint16 index = 0; index < m_valuesCount; ++index)
+        {
+            m_changedValues[index] = false;
+        }
+    }
+
+    if (m_objectUpdated)
+    {
+        if (remove)
+        {
+            RemoveFromClientUpdateList();
+        }
+        m_objectUpdated = false;
+    }
+}
+
+/**
+ * @brief Load values from data string
+ * @param data Data string to load from
+ * @return True if successful
+ *
+ * Loads update field values from a character data string.
+ * Used when loading objects from database.
+ */
+bool Object::LoadValues(const char* data)
+{
+    if (!m_uint32Values)
+    {
+        _InitValues();
+    }
+
+    Tokens tokens = StrSplit(data, " ");
+
+    if (tokens.size() != m_valuesCount)
+    {
+        return false;
+    }
+
+    Tokens::iterator iter;
+    int index;
+    for (iter = tokens.begin(), index = 0; index < m_valuesCount; ++iter, ++index)
+    {
+        m_uint32Values[index] = std::stoul((*iter).c_str());
+    }
+
+    return true;
+}
+
+/**
+ * @brief Set update bits in mask
+ * @param updateMask Update mask to modify
+ * @param target Target player (unused)
+ *
+ * Sets bits in the update mask for all fields that have changed.
+ */
+void Object::_SetUpdateBits(UpdateMask* updateMask, Player* target) const
+{
+    uint32 valuesCount = m_valuesCount;
+    if (GetTypeId() == TYPEID_PLAYER && target != this)
+    {
+        valuesCount = PLAYER_END_NOT_SELF;
+    }
+
+    for (uint16 index = 0; index < valuesCount; ++index )
+        if (m_changedValues[index])
+        {
+            updateMask->SetBit(index);
+        }
+}
+
+/**
+ * @brief Set create bits in mask
+ * @param updateMask Update mask to modify
+ * @param target Target player (unused)
+ *
+ * Sets bits in the update mask for all non-zero fields.
+ * Used when creating a new object for a player.
+ */
+void Object::_SetCreateBits(UpdateMask* updateMask, Player* target) const
+{
+    uint32 valuesCount = m_valuesCount;
+    if (GetTypeId() == TYPEID_PLAYER && target != this)
+    {
+        valuesCount = PLAYER_END_NOT_SELF;
+    }
+
+    for (uint16 index = 0; index < valuesCount; ++index)
+        if (GetUInt32Value(index) != 0)
+        {
+            updateMask->SetBit(index);
+        }
+}
